% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frosting.R
\name{add_frosting}
\alias{add_frosting}
\alias{remove_frosting}
\title{Add frosting to a workflow}
\usage{
add_frosting(x, frosting, ...)

remove_frosting(x)
}
\arguments{
\item{x}{A workflow}

\item{frosting}{A frosting layer created using \code{frosting()}}

\item{...}{Not used.}
}
\value{
\code{x}, updated with a new or removed frosting postprocessor
}
\description{
Add frosting to a workflow
}
\examples{
jhu <- case_death_rate_subset \%>\%
  filter(time_value > "2021-11-01", geo_value \%in\% c("ak", "ca", "ny"))
r <- epi_recipe(jhu) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_naomit()

wf <- epi_workflow(r, parsnip::linear_reg()) \%>\% fit(jhu)
latest <- jhu \%>\%
  dplyr::filter(time_value >= max(time_value) - 14)

# Add frosting to a workflow and predict
f <- frosting() \%>\%
  layer_predict() \%>\%
  layer_naomit(.pred)
wf1 <- wf \%>\% add_frosting(f)
p1 <- predict(wf1, latest)
p1

# Remove frosting from the workflow and predict
wf2 <- wf1 \%>\% remove_frosting()
p2 <- predict(wf2, latest)
p2
}
