% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_layer}
\alias{add_layer}
\title{Add layer to a frosting object}
\usage{
add_layer(frosting, object, flag = TRUE)
}
\arguments{
\item{frosting}{a \code{frosting} postprocessor}

\item{object}{a \code{frosting} layer}

\item{flag}{logical to determine if the layer is added. Default \code{TRUE}.}
}
\value{
an updated \code{frosting} postprocessor
}
\description{
Add layer to a frosting object
}
