% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_shift.R
\name{add_shifted_columns}
\alias{add_shifted_columns}
\title{backend for both \code{bake.step_epi_ahead} and \code{bake.step_epi_lag}, performs the
checks missing in \code{epi_shift_single}}
\usage{
add_shifted_columns(new_data, object, amount)
}
\description{
backend for both \code{bake.step_epi_ahead} and \code{bake.step_epi_lag}, performs the
checks missing in \code{epi_shift_single}
}
\keyword{internal}
