% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_recipe.R
\name{adjust_epi_recipe}
\alias{adjust_epi_recipe}
\alias{adjust_epi_recipe.epi_workflow}
\alias{adjust_epi_recipe.epi_recipe}
\title{Adjust a step in an \code{epi_workflow} or \code{epi_recipe}}
\usage{
adjust_epi_recipe(x, step_num, ..., blueprint = default_epi_recipe_blueprint())

\method{adjust_epi_recipe}{epi_workflow}(x, step_num, ..., blueprint = default_epi_recipe_blueprint())

\method{adjust_epi_recipe}{epi_recipe}(x, step_num, ..., blueprint = default_epi_recipe_blueprint())
}
\arguments{
\item{x}{A \code{epi_workflow} or \code{epi_recipe} object}

\item{step_num}{the number of the step to adjust}

\item{...}{Used to input a parameter adjustment}

\item{blueprint}{A hardhat blueprint used for fine tuning the preprocessing.}
}
\value{
\code{x}, updated with the adjustment to the specified \code{epi_recipe} step.
}
\description{
Make a parameter adjustment to a step in either an
\code{epi_workflow} or \code{epi_recipe} object.
}
\details{
This function can either adjust a step in a \code{epi_recipe} object
or a step from a \code{epi_recipe} object in an \code{epi_workflow}. In any case, the
argument name and update value must be inputted as \code{...}.
See the examples below for brief illustrations of both types of updates.
}
\examples{
jhu <- case_death_rate_subset \%>\%
  filter(time_value > "2021-11-01", geo_value \%in\% c("ak", "ca", "ny"))
r <- epi_recipe(jhu) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_naomit()

wf <- epi_workflow(r, parsnip::linear_reg()) \%>\% fit(jhu)
latest <- jhu \%>\%
  dplyr::filter(time_value >= max(time_value) - 14)

# Adjust `step_epi_ahead` to have an ahead value of 14
# in the `epi_workflow`
wf2 = wf \%>\% adjust_epi_recipe(step_num = 2, ahead = 14)
workflows::extract_preprocessor(wf2)

# Adjust `step_epi_ahead` to have an ahead value of 14
# in the `epi_recipe`
r2 = r \%>\% adjust_epi_recipe(step_num = 2, ahead = 14)
r2

}
