% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{adjust_name}
\alias{adjust_name}
\title{adjust the shifts by latency for the names in column assumes e.g.
\code{"lag_6_case_rate"} and returns something like \code{"lag_10_case_rate"}}
\usage{
adjust_name(prefix, column, effective_shift)
}
\description{
adjust the shifts by latency for the names in column assumes e.g.
\code{"lag_6_case_rate"} and returns something like \code{"lag_10_case_rate"}
}
\keyword{internal}
