% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_forecaster.R
\name{arx_args_list}
\alias{arx_args_list}
\title{ARX forecaster argument constructor}
\usage{
arx_args_list(
  lags = c(0L, 7L, 14L),
  ahead = 7L,
  min_train_window = 20L,
  forecast_date = NULL,
  target_date = NULL,
  levels = c(0.05, 0.95),
  symmetrize = TRUE,
  nonneg = TRUE,
  quantile_by_key = TRUE
)
}
\arguments{
\item{lags}{Vector or List. Positive integers enumerating lags to use
in autoregressive-type models (in days).}

\item{ahead}{Integer. Number of time steps ahead (in days) of the forecast
date for which forecasts should be produced.}

\item{min_train_window}{Integer. The minimal amount of training
data (in the time unit of the \code{epi_df}) needed to produce a forecast.
If smaller, the forecaster will return \code{NA} predictions.}

\item{forecast_date}{The date on which the forecast is created. The default
\code{NULL} will attempt to determine this automatically.}

\item{target_date}{The date for which the forecast is intended. The default
\code{NULL} will attempt to determine this automatically.}

\item{levels}{Vector or \code{NULL}. A vector of probabilities to produce
prediction intervals. These are created by computing the quantiles of
training residuals. A \code{NULL} value will result in point forecasts only.}

\item{symmetrize}{Logical. The default \code{TRUE} calculates
symmetric prediction intervals.}

\item{nonneg}{Logical. The default \code{TRUE} enforeces nonnegative predictions
by hard-thresholding at 0.}

\item{quantile_by_key}{Should separate quantiles be produced for each
\code{geo_value} (and additional \code{keys}) combination}
}
\value{
A list containing updated parameter choices.
}
\description{
Constructs a list of arguments for \code{\link[=arx_forecaster]{arx_forecaster()}}.
}
\examples{
arx_args_list()
arx_args_list(symmetrize = FALSE)
arx_args_list(levels = c(.1, .3, .7, .9), min_train_window = 120)
}
