% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_forecaster.R
\name{arx_forecaster}
\alias{arx_forecaster}
\title{AR forecaster with optional covariates}
\usage{
arx_forecaster(x, y, key_vars, time_value, args = arx_args_list())
}
\arguments{
\item{x}{Covariates. Allowed to be missing (resulting in AR on \code{y}).}

\item{y}{Response.}

\item{key_vars}{Factor(s). A prediction will be made for each unique
combination.}

\item{time_value}{the time value associated with each row of measurements.}

\item{args}{Additional arguments specifying the forecasting task. Created
by calling \code{arx_args_list()}.}
}
\value{
A data frame of point (and optionally interval) forecasts at a single
ahead (unique horizon) for each unique combination of \code{key_vars}.
}
\description{
AR forecaster with optional covariates
}
