% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_classifier.R
\name{arxc_epi_workflow_template}
\alias{arxc_epi_workflow_template}
\title{Create a template \code{arx_classifier} workflow}
\usage{
arxc_epi_workflow_template(
  epi_data,
  outcome,
  predictors,
  trainer = NULL,
  args_list = arx_class_args_list()
)
}
\arguments{
\item{epi_data}{An \code{epi_df} object}

\item{outcome}{A character (scalar) specifying the outcome (in the
\code{epi_df}).}

\item{predictors}{A character vector giving column(s) of predictor
variables.}

\item{trainer}{A \code{{parsnip}} model describing the type of estimation.
For now, we enforce \code{mode = "classification"}. Typical values are
\code{\link[parsnip:logistic_reg]{parsnip::logistic_reg()}} or \code{\link[parsnip:multinom_reg]{parsnip::multinom_reg()}}. More complicated
trainers like \code{\link[parsnip:naive_Bayes]{parsnip::naive_Bayes()}} or \code{\link[parsnip:rand_forest]{parsnip::rand_forest()}} can
also be used. May be \code{NULL} (the default).}

\item{args_list}{A list of customization arguments to determine
the type of forecasting model. See \code{\link[=arx_args_list]{arx_args_list()}}.}
}
\value{
An unfit \code{epi_workflow}.
}
\description{
This function creates an unfit workflow for use with \code{\link[=arx_classifier]{arx_classifier()}}.
It is useful if you want to make small modifications to that classifier
before fitting and predicting. Supplying a trainer to the function
may alter the returned \code{epi_workflow} object but can be omitted.
}
\examples{

jhu <- case_death_rate_subset \%>\%
  dplyr::filter(time_value >= as.Date("2021-11-01"))

arxc_epi_workflow_template(jhu, "death_rate", c("case_rate", "death_rate"))

arxc_epi_workflow_template(
  jhu,
  "death_rate",
  c("case_rate", "death_rate"),
  trainer = parsnip::multinom_reg(),
  args_list = arx_class_args_list(
    breaks = c(-.05, .1), ahead = 14,
    horizon = 14, method = "linear_reg"
  )
)
}
\seealso{
\code{\link[=arx_classifier]{arx_classifier()}}
}
