% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_arg_list.R
\name{assign_arg_list}
\alias{assign_arg_list}
\title{Assign argument list to inside an environment}
\usage{
assign_arg_list(l, env = parent.frame())
}
\arguments{
\item{l}{List of named arguments.}

\item{env}{The environment where the args should be assigned.
The default goes into the calling environment.}
}
\value{
Nothing is returned. Called for the side effects.
}
\description{
This function is similar to \code{attach()} but without the
need to detach. Calling it at the beginning of a forecaster
makes all members of the \code{arg_list} available inside the
forecaster with out the ugly \code{args$member} syntax.
}
\examples{
\dontrun{
  rm(list = ls())
  l <- list(a=1, b=c(12, 10), ff = function() -5)
  assign_arg_list(l)
  a
}
}
