% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.epi_workflow}
\alias{autoplot.epi_workflow}
\title{Automatically plot an \code{epi_workflow}}
\usage{
\method{autoplot}{epi_workflow}(
  object,
  predictions = NULL,
  .levels = c(0.5, 0.8, 0.95),
  ...,
  .color_by = c("all_keys", "geo_value", "other_keys", ".response", "all", "none"),
  .facet_by = c(".response", "other_keys", "all_keys", "geo_value", "all", "none"),
  .base_color = "#3A448F",
  .pred_line_color = "orange",
  .max_facets = Inf
)
}
\arguments{
\item{object}{An \code{epi_workflow}}

\item{predictions}{A data frame with predictions. If \code{NULL}, only the
original data is shown.}

\item{.levels}{A numeric vector of levels to plot for any prediction bands.
More than 3 levels begins to be difficult to see.}

\item{...}{Ignored}

\item{.color_by}{Which variables should determine the color(s) used to plot
lines. Options include:
\itemize{
\item \code{all_keys} - the default uses the interaction of any key variables
including the \code{geo_value}
\item \code{geo_value} - \code{geo_value} only
\item \code{other_keys} - any available keys that are not \code{geo_value}
\item \code{.response} - the numeric variables (same as the y-axis)
\item \code{all} - uses the interaction of all keys and numeric variables
\item \code{none} - no coloring aesthetic is applied
}}

\item{.facet_by}{Similar to \code{.color_by} except that the default is to display
each numeric variable on a separate facet}

\item{.base_color}{If available, prediction bands will be shown with this
color.}

\item{.pred_line_color}{If available, point forecasts will be shown with this
color.}

\item{.max_facets}{Cut down of the number of facets displayed. Especially
useful for testing when there are many \code{geo_value}'s or keys.}
}
\description{
For an unfit workflow, the training data will be displayed, the response by
default. If \code{predictions} is not \code{NULL} then point and interval forecasts
will be shown as well.
}
