% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_adjust_latency.R
\name{bake.step_adjust_latency}
\alias{bake.step_adjust_latency}
\title{various ways of handling differences between the \code{as_of} date and the maximum
time value}
\usage{
\method{bake}{step_adjust_latency}(object, new_data, ...)
}
\arguments{
\item{new_data}{assumes that this already has lag/ahead columns that we need
to adjust}
}
\description{
adjust the ahead so that we will be predicting \code{ahead} days after the \code{as_of}
date, rather than relative to the last day of data
}
