% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_enough_data.R
\name{check_enough_data}
\alias{check_enough_data}
\title{Check the dataset contains enough data points.}
\usage{
check_enough_data(
  recipe,
  ...,
  min_observations = NULL,
  epi_keys = NULL,
  drop_na = TRUE,
  role = NA,
  trained = FALSE,
  skip = TRUE,
  id = rand_id("enough_data")
)
}
\arguments{
\item{recipe}{A recipe object. The check will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this check.
See \code{\link[recipes:selections]{recipes::selections()}} for more details. You will usually want to use
\code{\link[recipes:has_role]{recipes::all_predictors()}} and/or \code{\link[recipes:has_role]{recipes::all_outcomes()}} here.}

\item{min_observations}{The minimum number of data points required for
training. If this is NULL, the total number of predictors will be used.}

\item{epi_keys}{A character vector of column names on which to group the data
and check threshold within each group. Useful if your forecaster trains
per group (for example, per geo_value).}

\item{drop_na}{A logical for whether to count NA values as valid rows.}

\item{role}{Not used by this check since no new variables are
created.}

\item{trained}{A logical for whether the selectors in \code{...}
have been resolved by \code{\link[=prep]{prep()}}.}

\item{skip}{A logical. If \code{TRUE}, only training data is checked, while if
\code{FALSE}, both training and predicting data is checked. Technically, this
answers the question "should the check be skipped when the recipe is baked
by \code{\link[=bake]{bake()}}?" While all operations are baked when \code{\link[=prep]{prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)).  Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this check to identify it.}
}
\description{
\code{check_enough_data} creates a \emph{specification} of a recipe
operation that will check if variables contain enough data.
}
\details{
This check will break the \code{prep} and/or bake function if any of the
checked columns have not enough non-NA values. If the check passes, nothing
is changed in the data. It is best used after every other step.

For checking training data, it is best to set \code{...} to be
\verb{all_predictors(), all_outcomes()}, while for checking prediction data, it
is best to set \code{...} to be \code{all_predictors()} only, with \code{n = 1}.
}
\section{tidy() results}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this check, a tibble with column
\code{terms} (the selectors or variables selected) is returned.
}

\concept{checks}
