% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{check_interminable_latency}
\alias{check_interminable_latency}
\title{warn when the latency is larger than would be reasonable}
\usage{
check_interminable_latency(
  dataset,
  latency_table,
  target_columns,
  forecast_date,
  call = caller_env()
)
}
\arguments{
\item{dataset}{the epi_df}

\item{latency_table}{the whole collection of latencies}

\item{target_columns}{the names of the columns that we're adjusting, and whether its unreasonably latent}
}
\description{
warn when the latency is larger than would be reasonable
}
\keyword{internal}
