% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{check_pname}
\alias{check_pname}
\title{Check that newly created variable names don't overlap}
\usage{
check_pname(res, preds, object, newname = NULL)
}
\arguments{
\item{res}{A data frame or tibble of the newly created variables.}

\item{preds}{An epi_df or tibble containing predictions.}

\item{object}{A layer object passed to \code{\link[=slather]{slather()}}.}

\item{newname}{A string of variable names if the object doesn't contain a
$name element}
}
\description{
\code{check_pname} is to be used in a slather method to ensure that
newly created variable names don't overlap with existing names.
Throws an warning if check fails, and creates a random string.
}
\keyword{internal}
