% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_epi_slide.R
\name{clean_f_name}
\alias{clean_f_name}
\title{Create short function names}
\usage{
clean_f_name(.f, max_length = 20L)
}
\arguments{
\item{.f}{a function, character string, or lambda. For example, \code{mean},
\code{"mean"}, \code{~ mean(.x)} or \verb{\\(x) mean(x, na.rm = TRUE)}.}

\item{max_length}{integer determining how long names can be}
}
\value{
a character string of length at most \code{max_length} that
(partially) describes the function.
}
\description{
Create short function names
}
\examples{
clean_f_name(mean)
clean_f_name("mean")
clean_f_name(~ mean(.x, na.rm = TRUE))
clean_f_name(\(x) mean(x, na.rm = TRUE))
clean_f_name(function(x) mean(x, na.rm = TRUE, trim = 0.2357862))
}
