% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{construct_shift_tibble}
\alias{construct_shift_tibble}
\title{create a table of the columns to modify, their shifts, and their prefixes}
\usage{
construct_shift_tibble(terms_used, recipe, rel_step_type, shift_name)
}
\description{
create a table of the columns to modify, their shifts, and their prefixes
}
\keyword{internal}
