% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-epidatasets.R
\docType{data}
\name{county_smoothed_cli_comparison}
\alias{county_smoothed_cli_comparison}
\title{Percent CLI from different surveys, compared to ground truth COVID incidence in a subset of counties}
\format{
An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 63840 rows and 5 columns.
}
\source{
This object contains a modified part of the \href{https://github.com/CSSEGISandData/COVID-19}{COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University} as \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{republished in the COVIDcast Epidata API}. This data set is licensed under the terms of the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International license}
by Johns Hopkins University on behalf of its Center for Systems Science in Engineering.
Copyright Johns Hopkins University 2020.

Modifications:
\itemize{
\item \href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}{From the COVIDcast Epidata API}:  The signal \code{confirmed_cumulative_num} was used to determine eligibility for inclusion. The signal \code{confirmed_7dav_incidence_prop} was computed by Delphi from the original JHU-CSSE data by calculating moving averages of the preceding 7 days, so the signal for June 7 is the average of the underlying data for June 1 through 7, inclusive.
\item Furthermore, the data has been limited to a specific time range, the
signal names slightly altered, and formatted into an \code{epi_df}.
}

This object contains a modified part of the
\href{https://cmu-delphi.github.io/delphi-epidata/symptom-survey/#covid-19-trends-and-impact-survey}{data
aggregations in the API} that are prepared from the
\href{https://www.pnas.org/doi/full/10.1073/pnas.2111454118}{COVID-19
Trends and Impact Survey}; see the first link for more information on
citing in publications.
The data is made available via the
\href{https://cmu-delphi.github.io/delphi-epidata/}{Delphi Epidata API}.

These aggregations are licensed under the terms of
the \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons
Attribution license}.

Modifications:
\itemize{
\item The data has been limited to a very small number of rows, the
signal names slightly altered, and formatted into an \code{epi_df}.
}

This object contains a modified part of the
\href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/google-survey.html}{Google symptom surveys}.
Aggregations based on the survey are licensed under the terms of
the \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons
Attribution license}.

Modifications:
\itemize{
\item The data has been limited to a very small number of rows, the
signal names slightly altered, and formatted into an \code{epi_df}.
}
}
\usage{
county_smoothed_cli_comparison
}
\description{
Data set for more than 400 US counties containing CLI
(COVID-19-like-illness) incidence derived from two surveys, and a reference signal as
reported by JHU CSSE. This example data is a snapshot as of September 21,
2020, and ranges from April 11, 2020 to September 01, 2020.

The reference signal \code{case} is based on reports made available
by the Center for Systems Science and Engineering at Johns Hopkins
University.

One survey was
\href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/google-survey.html}{run by Google},
in partnership with Delphi.

The other survey, the
\href{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/fb-survey.html}{COVID-19 Trends and Impact Survey},
was run by Delphi in collaboration with Facebook.

Data is reported for counties that had at least 200 cumulative COVID-19 cases
on May 14, 2020, according to JHU CSSE.
}
\section{Data dictionary}{


The data has columns:
\describe{
\item{geo_value}{The 5-digit county FIPS code associated with each
row of measurements.}
\item{time_value}{The time value, a date in YYYY-MM-DD format}
\item{goog}{Seven-day average of CLI (covid-like-illness) cases from the Google survey}
\item{fb}{Seven-day average of CLI (covid-like-illness) cases from CTIS}
\item{case}{Reference signal. Seven-day average of CLI (covid-like-illness) cases}
}

}

\examples{
# Since this is a re-exported dataset, it cannot be loaded using
# the `data()` function. `data()` looks for a file of the same name
# in the `data/` directory, which doesn't exist in this package.
# works
epipredict::county_smoothed_cli_comparison

# works
library(epipredict)
county_smoothed_cli_comparison

# fails
\dontrun{
  data(county_smoothed_cli_comparison, package = "epipredict")
}
}
\keyword{datasets}
