% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lags_and_leads.R
\name{create_lags_and_leads}
\alias{create_lags_and_leads}
\title{Create lags and leads of predictors and response}
\usage{
create_lags_and_leads(x, y, xy_lags, y_leads, time_value, key_vars = NULL)
}
\arguments{
\item{x}{Data frame or matrix. Predictor variables. May be
missing.}

\item{y}{Response vector. Typical usage will "lead" y by the
number of steps forward for the prediction horizon (ahead).}

\item{xy_lags}{Vector or list. If a vector, the lags will apply
to each column of \code{x} and to \code{y}. If a list, it must be of length
\code{ncol(x)+1} and each component will apply to the requisite predictor.
A \code{NULL} list element will remove that variable completely from the
result. Negative values will "lead" the variable.}

\item{y_leads}{Scalar or vector. If a scalar, we "lead" \code{y} by this
amount. A vector will produce multiple columns of \code{y} if this is
useful for your model. Negative values will "lag" the variable.}

\item{time_value}{Vector of time values at which the data are
observed}

\item{key_vars}{Factors representing different groups. May be
\code{NULL} (the default).}
}
\value{
A \code{data.frame}.
}
\description{
Create lags and leads of predictors and response
}
\examples{

x <- 1:20
y <- -20:-1
time_value <- c(1:18, 20, 21)
create_lags_and_leads(x, y, c(1, 2), 1, time_value)
create_lags_and_leads(x, y, list(c(1, 2), 1), 1, time_value)
create_lags_and_leads(x, y, list(c(-1, 1), NULL), 1, time_value)
create_lags_and_leads(x, y, c(1, 2), c(0, 1), time_value)
}
