% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-layer.R
\name{create_layer}
\alias{create_layer}
\title{Create a new layer}
\usage{
create_layer(name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{name}{Either a string giving a file name (without directory) or
\code{NULL} to take the name from the currently open file in RStudio.}

\item{open}{Whether to open the file for interactive editing.}
}
\description{
This function creates the skeleton for a new \code{frosting} layer. When called
inside a package, it will create an R script in the \verb{R/} directory,
fill in the name of the layer, and open the file.
}
\examples{
\dontrun{

  # Note: running this will write `layer_strawberry.R` to
  # the `R/` directory of your current project
  create_layer("strawberry")
}

}
