% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-epidatasets.R
\docType{data}
\name{ctis_covid_behaviours}
\alias{ctis_covid_behaviours}
\title{Subset of CTIS COVID-19-related behaviours from 5 states}
\format{
An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1055 rows and 4 columns.
}
\source{
This object contains a modified part of the
\href{https://cmu-delphi.github.io/delphi-epidata/symptom-survey/#covid-19-trends-and-impact-survey}{data
aggregations in the API} that are prepared from the
\href{https://www.pnas.org/doi/full/10.1073/pnas.2111454118}{COVID-19
Trends and Impact Survey}; see the first link for more information on
citing in publications.
The data is made available via the
\href{https://cmu-delphi.github.io/delphi-epidata/}{Delphi Epidata API}.

These aggregations are licensed under the terms of
the \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons
Attribution license}.

Modifications:
\itemize{
\item The data has been limited to a very small number of rows, the
signal names slightly altered, and formatted into an \code{epi_df}.
}
}
\usage{
ctis_covid_behaviours
}
\description{
Data set for a handful of states on masking and distancing behaviours
during the COVID-19 Pandemic, and downloaded from the CMU Delphi COVIDcast
Epidata API. This example data is a snapshot as of March 20, 2024, and
ranges from June 4, 2021 to December 31, 2021.
It is limited to California, Florida, Texas, New Jersey, and New York.
}
\section{Data dictionary}{


The data has columns:
\describe{
\item{geo_value}{the geographic value associated with each row
of measurements.}
\item{time_value}{the time value associated with each row of measurements.}
\item{masking}{Estimated percentage of people who wore a mask for most or all of the time while in public in the past 7 days; those not in public in the past 7 days are not counted.}
\item{distancing}{Estimated percentage of respondents who reported that all or most people they encountered in public in the past 7 days maintained a distance of at least 6 feet. Respondents who said that they have not been in public for the past 7 days are excluded.}
}

}

\examples{
# Since this is a re-exported dataset, it cannot be loaded using
# the `data()` function. `data()` looks for a file of the same name
# in the `data/` directory, which doesn't exist in this package.
# works
epipredict::ctis_covid_behaviours

# works
library(epipredict)
ctis_covid_behaviours

# fails
\dontrun{
data(ctis_covid_behaviours, package = "epipredict")
}
}
\keyword{datasets}
