% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mat_mul.R
\name{df_mat_mul}
\alias{df_mat_mul}
\title{Multiply columns of a \code{data.frame} by a matrix}
\usage{
df_mat_mul(dat, mat, out_names = "out", ...)
}
\arguments{
\item{dat}{A data.frame}

\item{mat}{A matrix}

\item{out_names}{Character vector. Creates the names of the resulting
columns after multiplication. If a scalar, this is treated as a
prefix and the remaining columns will be numbered sequentially.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\value{
A data.frame with the new columns at the right. Original
columns are removed.
}
\description{
Multiply columns of a \code{data.frame} by a matrix
}
\examples{
df <- data.frame(matrix(1:200, ncol = 10))
mat <- matrix(1:10, ncol = 2)
df_mat_mul(df, mat, "z", dplyr::num_range("X", 2:6))
}
