% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{drop_ignored_keys}
\alias{drop_ignored_keys}
\title{given a list named by key columns, remove any matching key values
keys_to_ignore should have the form list(col_name = c("value_to_ignore", "other_value_to_ignore"))}
\usage{
drop_ignored_keys(training, keys_to_ignore)
}
\description{
given a list named by key columns, remove any matching key values
keys_to_ignore should have the form list(col_name = c("value_to_ignore", "other_value_to_ignore"))
}
\keyword{internal}
