% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_shift.R
\name{epi_shift}
\alias{epi_shift}
\title{Shift predictors while maintaining grouping and time_value ordering}
\usage{
epi_shift(x, lags, time_value, keys = NULL, out_name = "x")
}
\arguments{
\item{x}{Data frame. Variables to lag}

\item{lags}{List. Each list element is a vector of lags.
Negative values produce leads. The list should have the same
length as the number of columns in \code{x}.}

\item{time_value}{Vector. Same length as \code{x} giving time stamps.}

\item{keys}{Data frame, vector, or \code{NULL}. Additional grouping vars.}

\item{out_name}{Chr. The output list will use this as a prefix.}
}
\value{
a list of tibbles
}
\description{
This is a lower-level function. As such it performs no error checking.
}
