% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_shift.R
\name{epi_shift_single}
\alias{epi_shift_single}
\title{Shift predictors while maintaining grouping and time_value ordering}
\usage{
epi_shift_single(x, col, shift_val, newname, key_cols)
}
\arguments{
\item{x}{Data frame.}

\item{shift_val}{a single integer. Negative values produce leads.}

\item{newname}{the name for the newly shifted column}

\item{key_cols}{vector, or \code{NULL}. Additional grouping vars.}
}
\value{
a list of tibbles
}
\description{
This is a lower-level function. As such it performs no error checking.
}
\keyword{internal}
