% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_epi_slide.R
\name{epi_slide_wrapper}
\alias{epi_slide_wrapper}
\title{Wrapper to handle epi_slide particulars}
\usage{
epi_slide_wrapper(
  new_data,
  before,
  after,
  columns,
  fns,
  fn_names,
  group_keys,
  name_prefix
)
}
\arguments{
\item{fns}{vector of functions, even if it's length 1.}

\item{group_keys}{the keys to group by. likely \code{epi_keys} (without \code{time_value})}
}
\description{
This should simplify somewhat in the future when we can run \code{epi_slide} on
columns. Surprisingly, lapply is several orders of magnitude faster than
using roughly equivalent tidy select style.
}
\keyword{internal}
