% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_pred-methods.R
\name{epipredict-vctrs}
\alias{epipredict-vctrs}
\alias{vec_arith.quantile_pred}
\title{Internal vctrs methods}
\usage{
\method{vec_arith}{quantile_pred}(op, x, y, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
