% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_climate.R
\name{epiweek_leap}
\alias{epiweek_leap}
\title{epiweek, but it assigns week 53 the value of 999 instead so it mirrors the assignments in yday_leap}
\usage{
epiweek_leap(time_value)
}
\description{
epiweek, but it assigns week 53 the value of 999 instead so it mirrors the assignments in yday_leap
}
\keyword{internal}
