% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{extend_either}
\alias{extend_either}
\title{offset each relevant column by it's appropriate latency
works for either adjusting aheads or lags
note that this may introduce new NA values when one column is shifted farther than another}
\usage{
extend_either(new_data, shift_cols, keys)
}
\arguments{
\item{new_data}{just what is says}

\item{shift_cols}{a tibble which must have the columns \code{column}, the name of
the column to adjust, \code{latency} the latency of the original column relative
to the \code{as_of} date, \code{new_name}, the names in \code{column} adjusted by the
latencies \code{latency}}

\item{keys}{the variables which are used as keys}
}
\description{
offset each relevant column by it's appropriate latency
works for either adjusting aheads or lags
note that this may introduce new NA values when one column is shifted farther than another
}
\keyword{internal}
