% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_argument}
\alias{extract_argument}
\title{Extract an argument made to a frosting layer or recipe step}
\usage{
extract_argument(x, name, arg, ...)
}
\arguments{
\item{x}{an epi_workflow, epi_recipe, frosting, step, or layer object}

\item{name}{the name of the layer}

\item{arg}{the name of the argument}

\item{...}{not used}
}
\value{
An object originally passed as an argument to a layer or step
}
\description{
Extract an argument made to a frosting layer or recipe step
}
\examples{
f <- frosting() \%>\%
  layer_predict() \%>\%
  layer_residual_quantiles(quantile_levels = c(0.0275, 0.975), symmetrize = FALSE) \%>\%
  layer_naomit(.pred)

extract_argument(f, "layer_residual_quantiles", "symmetrize")
}
\keyword{internal}
