% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frosting.R
\name{extract_frosting}
\alias{extract_frosting}
\title{Extract the frosting object from a workflow}
\usage{
extract_frosting(x, ...)
}
\arguments{
\item{x}{an \code{epi_workflow} object}

\item{...}{not used}
}
\value{
a \code{frosting} object
}
\description{
Extract the frosting object from a workflow
}
