% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{extract_layer_argument}
\alias{extract_layer_argument}
\title{Extract argument made to a frosting layer}
\usage{
extract_layer_argument(x, layer_name, arg, ...)
}
\arguments{
\item{x}{an epi_workflow, frosting, or layer object}

\item{layer_name}{the name of the layer}

\item{arg}{the name of the argument}

\item{...}{not used}
}
\value{
An object originally passed as an argument to a layer
}
\description{
Extract argument made to a frosting layer
}
\examples{
f <- frosting() \%>\%
  layer_predict() \%>\%
  layer_residual_quantile(probs = c(0.0275, 0.975), symmetrize = FALSE) \%>\%
  layer_naomit(.pred)

extract_layer_argument(f, "layer_residual_quantile", "symmetrize")
}
