% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_workflow.R
\name{fit-epi_workflow}
\alias{fit-epi_workflow}
\alias{fit.epi_workflow}
\title{Fit an \code{epi_workflow} object}
\usage{
\method{fit}{epi_workflow}(object, data, ..., control = workflows::control_workflow())
}
\arguments{
\item{object}{an \code{epi_workflow} object}

\item{data}{an \code{epi_df} of predictors and outcomes to use when
fitting the \code{epi_workflow}}

\item{...}{Not used}

\item{control}{A \code{\link[workflows:control_workflow]{workflows::control_workflow()}} object}
}
\value{
The \code{epi_workflow} object, updated with a fit parsnip
model in the \code{object$fit$fit} slot.
}
\description{
This is the \code{fit()} method for an \code{epi_workflow} object that
estimates parameters for a given model from a set of data.
Fitting an \code{epi_workflow} involves two main steps, which are
preprocessing the data and fitting the underlying parsnip model.
}
\examples{
jhu <- case_death_rate_subset \%>\%
  filter(time_value > "2021-11-01", geo_value \%in\% c("ak", "ca", "ny"))

r <- epi_recipe(jhu) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_ahead(death_rate, ahead = 7)

wf <- epi_workflow(r, parsnip::linear_reg()) \%>\% fit(jhu)
wf

}
\seealso{
workflows::fit-workflow
}
