% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_workflow.R
\name{forecast}
\alias{forecast}
\title{Produce a forecast from an epi workflow}
\usage{
forecast(
  epi_workflow,
  fill_locf = FALSE,
  n_recent = NULL,
  forecast_date = NULL
)
}
\arguments{
\item{epi_workflow}{An epi workflow}

\item{fill_locf}{Logical. Should we use locf to fill in missing data?}

\item{n_recent}{Integer or NULL. If filling missing data with locf = TRUE,
how far back are we willing to tolerate missing data? Larger values allow
more filling. The default NULL will determine this from the the recipe. For
example, suppose n_recent = 3, then if the 3 most recent observations in any
geo_value are all NA’s, we won’t be able to fill anything, and an error
message will be thrown. (See details.)}

\item{forecast_date}{By default, this is set to the maximum time_value in x.
But if there is data latency such that recent NA's should be filled, this may
be after the last available time_value.}
}
\value{
A forecast tibble.
}
\description{
Produce a forecast from an epi workflow
}
