% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frosting.R
\name{frosting}
\alias{frosting}
\title{Create frosting for postprocessing predictions}
\usage{
frosting(layers = NULL, requirements = NULL)
}
\arguments{
\item{layers}{Must be \code{NULL}.}

\item{requirements}{Must be \code{NULL}.}
}
\value{
A frosting object.
}
\description{
This generates a postprocessing container (much like \code{recipes::recipe()})
to hold steps for postprocessing predictions.
}
\details{
The arguments are currently placeholders and must be NULL
}
