% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_asof}
\alias{get_asof}
\title{extract the as_of, and make sure there's nothing very off about it}
\usage{
get_asof(object, new_data)
}
\description{
extract the as_of, and make sure there's nothing very off about it
}
