% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_forecast_date}
\alias{get_forecast_date}
\title{Extract the as_of for the forecast date, and make sure there's nothing very off about it.}
\usage{
get_forecast_date(new_data, info, epi_keys_checked, latency, columns = NULL)
}
\description{
Extract the as_of for the forecast date, and make sure there's nothing very off about it.
}
\keyword{internal}
