% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_forecast_date_in_layer}
\alias{get_forecast_date_in_layer}
\title{get the target date while in a layer}
\usage{
get_forecast_date_in_layer(this_recipe, workflow_max_time_value, new_data)
}
\description{
get the target date while in a layer
}
