% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_grouping_columns}
\alias{get_grouping_columns}
\title{return the names of the grouped columns, or \code{NULL}}
\usage{
get_grouping_columns(x)
}
\arguments{
\item{x}{an epi_df}
}
\description{
return the names of the grouped columns, or \code{NULL}
}
\keyword{internal}
