% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_latency}
\alias{get_latency}
\title{the latency is also the amount the shift is off by}
\usage{
get_latency(new_data, as_of, column, shift_amount, sign_shift)
}
\arguments{
\item{sign_shift}{integer. 1 if lag and -1 if ahead. These represent how you
need to shift the data to bring the 3 day lagged value to today.}
}
\description{
the latency is also the amount the shift is off by
}
\keyword{internal}
