% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_yeo_johnson.R
\name{get_params_in_layer}
\alias{get_params_in_layer}
\title{get the parameters used in the initial step}
\usage{
get_params_in_layer(
  workflow,
  step_name = "epi_YeoJohnson",
  param_name = "yj_params"
)
}
\arguments{
\item{workflow}{the workflow to extract the parameters from}

\item{step_name}{the name of the step to look for, as recognized by \code{detect_step}}

\item{param_name}{the parameter to pull out of the step}
}
\description{
get the parameters used in the initial step
}
\keyword{internal}
