% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_shifted_column_tibble}
\alias{get_shifted_column_tibble}
\title{find the columns added with the lags or aheads, and the amounts they have
been changed}
\usage{
get_shifted_column_tibble(
  prefix,
  new_data,
  terms_used,
  as_of,
  latency,
  sign_shift,
  info,
  call = caller_env()
)
}
\arguments{
\item{prefix}{the prefix indicating if we are adjusting lags or aheads}

\item{new_data}{the data transformed so far}
}
\value{
a tibble with columns \code{column} (relevant shifted names), \code{shift} (the
amount that one is shifted), \code{latency} (original columns difference between
max_time_value and as_of (on a per-initial column basis)),
\code{effective_shift} (shifts+latency), and \code{new_name} (adjusted names with the
effective_shift)
}
\description{
find the columns added with the lags or aheads, and the amounts they have
been changed
}
\keyword{internal}
