% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_shift.R
\name{get_sign}
\alias{get_sign}
\title{lags move columns forward to bring the past up to today, while aheads drag
the future back to today}
\usage{
get_sign(object)
}
\description{
lags move columns forward to bring the past up to today, while aheads drag
the future back to today
}
\keyword{internal}
