% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_names.R
\name{grab_names}
\alias{grab_names}
\title{Get the names from a data frame via tidy select}
\usage{
grab_names(dat, ...)
}
\arguments{
\item{dat}{a data.frame}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\value{
a character vector
}
\description{
Given a data.frame, use \verb{<tidy-select>} syntax to choose
some variables. Return the names of those variables
}
\details{
As this is an internal function, no checks are performed.
}
\examples{
df <- data.frame(a = 1, b = 2, cc = rep(NA, 3))
grab_names(df, dplyr::starts_with("c"))
}
\keyword{internal}
