% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grf_quantiles.R
\name{grf_quantiles}
\alias{grf_quantiles}
\title{Random quantile forests via grf}
\description{
\code{\link[grf:quantile_forest]{grf::quantile_forest()}} fits random forests in a way that makes it easy
to calculate \emph{quantile} forests. Currently, this is the only engine
provided here, since quantile regression is the typical use-case.
}
\section{Tuning Parameters}{


This model has 3 tuning parameters:
\itemize{
\item \code{mtry}: # Randomly Selected Predictors (type: integer, default: see below)
\item \code{trees}: # Trees (type: integer, default: 2000L)
\item \code{min_n}: Minimal Node Size (type: integer, default: 5)
}

\code{mtry} depends on the number of columns in the design matrix.
The default in \code{\link[grf:quantile_forest]{grf::quantile_forest()}} is \code{min(ceiling(sqrt(ncol(X)) + 20), ncol(X))}.

For categorical predictors, a one-hot encoding is always used. This makes
splitting efficient, but has implications for the \code{mtry} choice. A factor
with many levels will become a large number of columns in the design matrix
which means that some of these may be selected frequently for potential splits.
This is different than in other implementations of random forest. For more
details, see \href{https://grf-labs.github.io/grf/articles/categorical_inputs.html}{the \code{grf} discussion}.
}

\section{Translation from parsnip to the original package}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{rand_forest(
  mode = "regression", # you must specify the mode to be regression
  mtry = integer(1),
  trees = integer(1),
  min_n = integer(1)
) \%>\%
  set_engine("grf_quantiles") \%>\%
  translate()
#> Random Forest Model Specification (regression)
#> 
#> Main Arguments:
#>   mtry = integer(1)
#>   trees = integer(1)
#>   min_n = integer(1)
#> 
#> Computational engine: grf_quantiles 
#> 
#> Model fit template:
#> grf::quantile_forest(X = missing_arg(), Y = missing_arg(), mtry = min_cols(~integer(1), 
#>     x), num.trees = integer(1), min.node.size = min_rows(~integer(1), 
#>     x), quantiles = c(0.1, 0.5, 0.9), num.threads = 1L, seed = runif(1, 
#>     0, .Machine$integer.max))
}\if{html}{\out{</div>}}
}

\section{Case weights}{


Case weights are not supported.
}

\examples{
y <- 1:3
x <- 2:4

}
