% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_recipe.R
\name{is_epi_recipe}
\alias{is_epi_recipe}
\title{Test for \code{epi_recipe}}
\usage{
is_epi_recipe(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from \code{epi_recipe}.
}
\description{
Test for \code{epi_recipe}
}
\keyword{internal}
