% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_workflow.R
\name{is_epi_workflow}
\alias{is_epi_workflow}
\title{Test for an \code{epi_workflow}}
\usage{
is_epi_workflow(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from \code{epi_workflow}.
}
\description{
Test for an \code{epi_workflow}
}
