% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_climate.R
\name{isoweek_leap}
\alias{isoweek_leap}
\title{isoweek, but it assigns week 53 the value of 999 instead so it mirrors the assignments in yday_leap}
\usage{
isoweek_leap(time_value)
}
\description{
isoweek, but it assigns week 53 the value of 999 instead so it mirrors the assignments in yday_leap
}
\keyword{internal}
