% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_iterative_ar_forecaster.R
\name{knn_iteraive_ar_forecaster}
\alias{knn_iteraive_ar_forecaster}
\title{KNN enhanced iterative AR forecaster with optional covariates}
\usage{
knn_iteraive_ar_forecaster(
  x,
  y,
  key_vars,
  time_value,
  args = knn_iteraive_ar_args_list()
)
}
\arguments{
\item{x}{Unused covariates. Must to be missing (resulting in AR on \code{y}) .}

\item{y}{Response.}

\item{key_vars}{Factor(s). A prediction will be made for each unique
combination.}

\item{time_value}{the time value associated with each row of measurements.}

\item{args}{Additional arguments specifying the forecasting task. Created
by calling \code{knn_iteraive_ar_args_list()}.}
}
\value{
A data frame of point (and optionally interval) forecasts at multiple
aheads (multiple horizons from one to specified \code{ahead}) for each unique combination of \code{key_vars}.
}
\description{
KNN enhanced iterative AR forecaster with optional covariates
}
