% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer}
\alias{layer}
\title{\code{layer} sets the class of the \code{layer}}
\usage{
layer(subclass, ..., .prefix = "layer_")
}
\arguments{
\item{subclass}{A character string for the resulting class. For example,
if \code{subclass = "blah"} the layer object that is returned has class
\code{layer_blah}.}

\item{...}{All arguments to the operator that should be returned.}

\item{.prefix}{Prefix to the subclass created.}
}
\value{
An updated layer with the new class
}
\description{
\code{layer} sets the class of the \code{layer}
}
\keyword{internal}
