% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_predictive_distn.R
\name{layer_predictive_distn}
\alias{layer_predictive_distn}
\title{Returns predictive distributions}
\usage{
layer_predictive_distn(
  frosting,
  ...,
  dist_type = c("gaussian", "student_t"),
  truncate = c(-Inf, Inf),
  name = ".pred_distn",
  id = rand_id("predictive_distn")
)
}
\arguments{
\item{frosting}{a \code{frosting} postprocessor}

\item{...}{Unused, include for consistency with other layers.}

\item{dist_type}{Gaussian or Student's t predictive intervals}

\item{truncate}{Do we truncate the distribution to an interval}

\item{name}{character. The name for the output column.}

\item{id}{a random id string}
}
\value{
an updated \code{frosting} postprocessor with additional columns of the
residual quantiles added to the prediction
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function calculates an \emph{approximation} to a parametric predictive
distribution. Predictive distributions from linear models require
\verb{x* (X'X)^\{-1\} x*}
along with the degrees of freedom. This function approximates both. It
should be reasonably accurate for models fit using \code{lm} when the new point
\verb{x*} isn't too far from the bulk of the data.
}
