% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_epi_slide_mean.R
\name{lubridate_period_to_integer}
\alias{lubridate_period_to_integer}
\title{lubridate converts to seconds by default, and as.integer doesn't throw errors if it isn't actually an integer}
\usage{
lubridate_period_to_integer(value, time_type)
}
\description{
lubridate converts to seconds by default, and as.integer doesn't throw errors if it isn't actually an integer
}
\keyword{internal}
