% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_quantiles.R
\name{nested_quantiles}
\alias{nested_quantiles}
\title{Turn a vector of quantile distributions into a list-col}
\usage{
nested_quantiles(x)
}
\arguments{
\item{x}{a \code{distribution} containing \code{dist_quantiles}}
}
\value{
a list-col
}
\description{
Turn a vector of quantile distributions into a list-col
}
\examples{
library(dplyr)
library(tidyr)
edf <- case_death_rate_subset[1:3, ]
edf$q <- dist_quantiles(list(1:5, 2:4, 3:10), list(1:5 / 6, 2:4 / 5, 3:10 / 11))

edf_nested <- edf \%>\% mutate(q = nested_quantiles(q))
edf_nested \%>\% unnest(q)
}
