% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{pad_to_end}
\alias{pad_to_end}
\title{pad every group at the right interval}
\usage{
pad_to_end(x, groups, end_date, columns_to_complete = NULL)
}
\arguments{
\item{x}{an epi_df to be filled forward.}

\item{groups}{the grouping by which to fill forward}

\item{columns_to_complete}{which columns to apply completion to. By default every non-key column of an epi_df}
}
\description{
Perform last observation carried forward on a group by group basis. It uses
\code{guess_period} to find the appropriate interval to fill-forward by. It
maintains the grouping structure it recieves. It does \emph{not} fill any
"interior" \code{NA} values occurring in the data beforehand.
}
\keyword{internal}
