% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_pred-methods.R
\name{quantile.quantile_pred}
\alias{quantile.quantile_pred}
\title{Quantiles from a distribution}
\usage{
\method{quantile}{quantile_pred}(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  lower = -Inf,
  upper = Inf,
  middle = c("cubic", "linear"),
  ...
)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}

\item{lower}{Scalar. Optional lower bound.}

\item{upper}{Scalar. Optional upper bound.}

\item{middle}{Controls how extrapolation to "interior" probabilities is
performed. "cubic" attempts to use \code{\link[stats:splinefun]{stats::splinefun()}} while "linear"
uses \code{\link[stats:approxfun]{stats::approx()}}. The "linear" method is used as a fallback if
"cubic" should fail for some reason.}

\item{...}{unused}
}
\value{
a matrix with one row for each entry in \code{x} and one column for each
value in \code{probs}
}
\description{
Given a \link[hardhat:quantile_pred]{hardhat::quantile_pred} object, users may wish to compute additional
\code{quantile_levels} that are not part of the object. This function attempts
to estimate these quantities under some assumptions. Interior probabilities,
those contained within, existing probabilities are interpolated in a manner
controled by the \code{middle} argument. Those outside existing probabilities
are extrapolated under the assumption that the tails of the distribution
decays exponentially. Optionally, one may constrain \emph{all} quantiles to be
within some support (say, \verb{[0, Inf)}).
}
\examples{
qp <- quantile_pred(matrix(1:8, nrow = 2, byrow = TRUE), 1:4 / 5)
quantile(qp)
quantile(qp, lower = 0)
quantile(qp, probs = 0.5)
quantile(qp, probs = 1:9 / 10)
}
\seealso{
\code{\link[=extrapolate_quantiles]{extrapolate_quantiles()}}
}
