% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_climate.R
\name{roll_modular_multivec}
\alias{roll_modular_multivec}
\title{group col by .idx values and sum windows around each .idx value}
\usage{
roll_modular_multivec(col, idx_in, weights, aggr, window_size, modulus)
}
\arguments{
\item{col}{the list of values indexed by \code{idx_in}}

\item{idx_in}{the relevant periodic part of time value, e.g. the week number,
limited to the relevant range}

\item{weights}{how much to weigh each particular datapoint (also indexed by
\code{idx_in})}

\item{aggr}{the aggregation function, probably Quantile, mean, or median}

\item{window_size}{the number of .idx entries before and after to include in
the aggregation}

\item{modulus}{the number of days/weeks/months in the year, not including any
leap days/weeks}
}
\description{
group col by .idx values and sum windows around each .idx value
}
\keyword{internal}
