% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{set_asof}
\alias{set_asof}
\title{extract the as_of, and make sure there's nothing very off about it}
\usage{
set_asof(new_data, info)
}
\description{
extract the as_of, and make sure there's nothing very off about it
}
\keyword{internal}
