% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_arx_forecaster.R
\name{smooth_arx_args_list}
\alias{smooth_arx_args_list}
\title{Smooth ARX forecaster argument constructor}
\usage{
smooth_arx_args_list(
  lags = c(0, 7, 14),
  ahead = 1:28,
  degree = 4,
  kronecker_version = FALSE,
  min_train_window = 20,
  levels = c(0.05, 0.95),
  intercept = TRUE,
  symmetrize = TRUE,
  nonneg = TRUE,
  quantile_by_key = FALSE
)
}
\arguments{
\item{lags}{Vector or List. Positive integers enumerating lags to use
in autoregressive-type models.}

\item{ahead}{Integer. Number of time steps ahead of the forecast date
for which forecasts should be produced.}

\item{degree}{Integer. Order of the orthodonal polynomials to use for
smoothing. Should be strictly less than \code{length(ahead)}.}

\item{kronecker_version}{Logical. Do we ensure that we've "seen" the latest
\code{ahead} value. The default \code{FALSE} is computationally simpler but uses
less recent data.}

\item{min_train_window}{Integer. The minimal amount of training
data needed to produce a forecast. If smaller, the forecaster will return
\code{NA} predictions.}

\item{levels}{Vector or \code{NULL}. A vector of probabilities to produce
prediction intervals. These are created by computing the quantiles of
training residuals. A \code{NULL} value will result in point forecasts only.}

\item{intercept}{Logical. The default \code{TRUE} includes intercept in the
forecaster.}

\item{symmetrize}{Logical. The default \code{TRUE} calculates symmetric prediction
intervals.}

\item{nonneg}{Logical. The default \code{TRUE} enforeces nonnegative predictions
by hard-thresholding at 0.}

\item{quantile_by_key}{Not currently implemented.}
}
\value{
A list containing updated parameter choices.
}
\description{
Constructs a list of arguments for \code{\link[=smooth_arx_forecaster]{smooth_arx_forecaster()}}.
}
\examples{
smooth_arx_args_list()
smooth_arx_args_list(symmetrize = FALSE)
smooth_arx_args_list(levels = c(.1, .3, .7, .9), min_train_window = 120)
}
