% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_arx_forecaster.R
\name{smooth_arx_forecaster}
\alias{smooth_arx_forecaster}
\title{Smooth AR forecaster with optional covariates}
\usage{
smooth_arx_forecaster(
  x,
  y,
  key_vars,
  time_value,
  args = smooth_arx_args_list()
)
}
\arguments{
\item{x}{Covariates. Allowed to be missing (resulting in AR on \code{y}).}

\item{y}{Response.}

\item{key_vars}{Factor(s). A prediction will be made for each unique
combination.}

\item{time_value}{the time value associated with each row of measurements.}

\item{args}{Additional arguments specifying the forecasting task. Created
by calling \code{smooth_arx_args_list()}.}
}
\value{
A data frame of point (and optionally interval) forecasts across
multiple aheads for each unique combination of \code{key_vars}.
}
\description{
Smooth AR forecaster with optional covariates
}
