% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_threshold_preds.R
\name{snap}
\alias{snap}
\title{restrict various objects to the interval [lower, upper]}
\usage{
snap(x, lower, upper, ...)
}
\arguments{
\item{x}{the object to restrict}

\item{lower}{numeric, the lower bound}

\item{upper}{numeric, the upper bound}

\item{...}{unused}
}
\description{
restrict various objects to the interval [lower, upper]
}
\keyword{internal}
