% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{step_adjust_latency_checks}
\alias{step_adjust_latency_checks}
\title{checks: the recipe type, whether a previous step is the relevant epi_shift,
that either \code{fixed_latency} or \code{fixed_forecast_date} is non-null, and that
\code{fixed_latency} only references columns that exist at the time of the step
inclusion}
\usage{
step_adjust_latency_checks(
  id,
  method,
  recipe,
  fixed_latency,
  fixed_forecast_date,
  call = caller_env()
)
}
\description{
checks: the recipe type, whether a previous step is the relevant epi_shift,
that either \code{fixed_latency} or \code{fixed_forecast_date} is non-null, and that
\code{fixed_latency} only references columns that exist at the time of the step
inclusion
}
\keyword{internal}
