% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_lag.R
\name{step_epi_lag}
\alias{step_epi_lag}
\title{Create a lagged predictor}
\usage{
step_epi_lag(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  lag = 1,
  prefix = "lag_",
  default = NA,
  keys = epi_keys(recipe),
  columns = NULL,
  skip = FALSE,
  id = rand_id("epi_lag")
)
}
\arguments{
\item{lag}{A vector of positive integers. Each specified column will be
lagged for each value in the vector.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_epi_lag} creates a \emph{specification} of a recipe step that
will add new columns of lagged data. Lagged data will
by default include NA values where the lag was induced.
These can be removed with \code{\link[=step_naomit]{step_naomit()}}, or you may
specify an alternative filler value with the \code{default}
argument.
}
\details{
The step assumes that the data are already \emph{in the proper sequential
order} for lagging.
}
\concept{row operation steps}
