% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_yeo_johnson.R
\name{step_epi_YeoJohnson}
\alias{step_epi_YeoJohnson}
\title{Yeo-Johnson transformation}
\usage{
step_epi_YeoJohnson(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  yj_params = NULL,
  na_fill = 1/4,
  limits = c(-5, 5),
  num_unique = 5,
  na_rm = TRUE,
  skip = FALSE,
  id = rand_id("epi_YeoJohnson")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? \code{lag} is default a predictor while \code{ahead} is an outcome.}

\item{trained}{A logical for whether the selectors in \code{...}
have been resolved by \code{\link[=prep]{prep()}}.}

\item{yj_params}{Internal. A numeric vector of transformation values. This
is \code{NULL} until computed by \code{\link[=prep]{prep()}}.}

\item{na_fill}{A numeric value to fill in for any geos where a Yeo-Johnson
parameter cannot be estimated.}

\item{limits}{A length 2 numeric vector defining the range to compute the
transformation parameter.}

\item{num_unique}{An integer where data that have fewer than this many unique
values will not be evaluated for a transformation.}

\item{na_rm}{A logical indicating whether missing values should be removed
before estimating the transformation parameter.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A unique identifier for the step}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_epi_YeoJohnson()} creates a \emph{specification} of a recipe step that will
transform data using a Yeo-Johnson transformation. This fork works with panel
data and is meant for epidata.
}
\details{
The Yeo-Johnson transformation is variance-stabilizing
transformation, similar to the Box-Cox but does not require the input
variables to be strictly positive. In the package, the partial
log-likelihood function is directly optimized within a reasonable set of
transformation values (which can be changed by the user). The optimization
finds a lambda parameter for each group in the data that minimizes the
variance of the transformed data.

This transformation is typically done on the outcome variable
using the residuals for a statistical model (such as ordinary
least squares). Here, a simple null model (intercept only) is
used to apply the transformation to the \emph{predictor}
variables individually. This can have the effect of making the
variable distributions more symmetric.

If the transformation parameters are estimated to be very
close to the bounds, or if the optimization fails, a value of
\code{NA} is used and no transformation is applied.
}
\section{Tidying}{
When you \code{\link[=tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms}, \code{value} , and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{value}{numeric, the lambda estimate}
\item{id}{character, id of this step}
}
}

\examples{
jhu <- cases_deaths_subset \%>\%
  filter(time_value > "2021-01-01", geo_value \%in\% c("ca", "ny")) \%>\%
  select(geo_value, time_value, cases)
filtered_data <- jhu

r <- epi_recipe(filtered_data) \%>\%
  step_epi_YeoJohnson(cases)
# View the recipe
r
# Fit the recipe
tr <- r \%>\% prep(filtered_data)
# View the parameter values
tr$steps[[1]]$yj_params
# View the transformed data
df <- tr \%>\% bake(filtered_data)
plot(density(df$cases))
plot(density(filtered_data$cases))
}
\references{
Yeo, I. K., and Johnson, R. A. (2000). A new family of power
transformations to improve normality or symmetry. \emph{Biometrika}.
}
\concept{individual transformation steps}
