% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_lag_difference.R
\name{step_lag_difference}
\alias{step_lag_difference}
\title{Calculate a lagged difference}
\usage{
step_lag_difference(
  recipe,
  ...,
  role = "predictor",
  horizon = 7,
  prefix = "lag_diff_",
  skip = FALSE,
  id = rand_id("lag_diff")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? \code{lag} is default a predictor while \code{ahead} is an outcome.}

\item{horizon}{Scalar or vector. Time period(s) over which to calculate
differences.}

\item{prefix}{A character string that will be prefixed to the new column.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A unique identifier for the step}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_lag_difference()} creates a \emph{specification} of a recipe step
that will generate one or more new columns of derived data.
}
\examples{
r <- recipe(case_death_rate_subset) \%>\%
  step_lag_difference(case_rate, death_rate, horizon = c(7, 14)) \%>\%
  step_epi_naomit()
r

r \%>\%
  prep(case_death_rate_subset) \%>\%
  bake(case_death_rate_subset)
}
\seealso{
Other row operation steps: 
\code{\link{step_epi_lag}()},
\code{\link{step_growth_rate}()}
}
\concept{row operation steps}
