% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_narm.R
\name{step_narm}
\alias{step_narm}
\title{Special NA omit step that does two steps in one}
\usage{
step_narm(x)
}
\arguments{
\item{x}{Recipe to be used for omission steps}
}
\value{
Omits NA's from both predictors and outcomes and training time;
however, only omits predictors at prediction time
}
\description{
Special NA omit step that does two steps in one
}
