% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.layer}
\alias{update.layer}
\title{Update a frosting layer}
\usage{
\method{update}{layer}(object, ...)
}
\arguments{
\item{object}{A frosting \code{layer}.}

\item{...}{Key-value pairs where the keys match up with names of elements
in the layer, and the values are the new values to update the layer with.}
}
\description{
This \code{layer} method for \code{update()} takes named arguments as \code{...} who's
values will replace the elements of the same name in the actual \code{layer}.
}
\examples{
f <- frosting() \%>\%
  layer_predict() \%>\%
  layer_naomit(.pred) \%>\%
  layer_add_forecast_date("2021-01-01")
f$layers[[2]]$id
f$layers[[3]]$forecast_date

update(f$layers[[2]], id = "kill_all_the_nas")
}
