% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_frosting_layers}
\alias{update_frosting_layers}
\title{Update frosting layers}
\usage{
update_frosting_layers(x, ...)
}
\arguments{
\item{x}{A frosting object}

\item{...}{named lists of layers to update with changes. Passed as
\code{layer_name = list(parm1 = updated_pars1, parm2 = updated_pars2)}.}
}
\value{
An updated frosting object
}
\description{
Given an existing frosting object, update some (or all) of the layers
with different arguments.
}
\details{
Note that not all layers or arguments can
be updated easily or at all. This function only accomodates named
arguments. Furthermore, frosting objects with multiple layers of the same
type can not be updated using this function.
}
\examples{

f <- frosting() \%>\%
  layer_predict() \%>\%
  layer_naomit(.pred) \%>\%
  layer_add_forecast_date("2021-01-01")
f$layers[[2]]$id
f$layers[[3]]$forecast_date

f2 <- update_frosting(f, layer_naomit = list(id = "kill_all_the_nas!"),
        layer_add_forecast_date = list(forecast_date = "2022-01-01"))
f2$layers[[2]]$id
f2$layers[[3]]$forecast_date
}
