% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{weighted_interval_score}
\alias{weighted_interval_score}
\title{Title}
\usage{
weighted_interval_score(x, actual, ...)
}
\arguments{
\item{x}{A vector of class \code{distribution}.}

\item{actual}{A vector of observed values}

\item{...}{additional named arguments. See methods.}
}
\value{
A vector of weighted interval scores
}
\description{
Title
}
\examples{
dstn <- dist_quantiles(list(1:4, 8:11), list(c(.2,.4,.6,.8)))
weighted_interval_score(dstn, c(2, 10.5))
}
