% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_climate.R
\name{within_window}
\alias{within_window}
\title{generate the idx values within \code{window_size} of \code{target_idx} given that our
time value is of the type matching modulus}
\usage{
within_window(target_idx, window_size, modulus)
}
\arguments{
\item{target_idx}{the time index which we're drawing the window around}

\item{window_size}{the size of the window on one side of \code{target_idx}}

\item{modulus}{the number of days/weeks/months in the year, not including any leap days/weeks}
}
\description{
generate the idx values within \code{window_size} of \code{target_idx} given that our
time value is of the type matching modulus
}
\keyword{internal}
