% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_climate.R
\name{yday_leap}
\alias{yday_leap}
\title{a function that assigns Feb 29th to 999, and aligns all other dates the same
number in the year, regardless of whether it's a leap year}
\usage{
yday_leap(time_value)
}
\description{
a function that assigns Feb 29th to 999, and aligns all other dates the same
number in the year, regardless of whether it's a leap year
}
\keyword{internal}
