% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frosting.R
\name{add_frosting}
\alias{add_frosting}
\alias{remove_frosting}
\title{Add frosting to a workflow}
\usage{
add_frosting(x, frosting, ...)

remove_frosting(x)
}
\arguments{
\item{x}{A workflow}

\item{frosting}{A frosting layer created using \code{frosting()}}

\item{...}{Not used.}
}
\value{
\code{x}, updated with a new or removed frosting postprocessor
}
\description{
Add frosting to a workflow
}
