% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frosting.R
\name{add_frosting}
\alias{add_frosting}
\alias{remove_frosting}
\title{Add frosting to a workflow}
\usage{
add_frosting(x, frosting, ...)

remove_frosting(x)
}
\arguments{
\item{x}{A workflow}

\item{frosting}{A frosting layer created using \code{frosting()}}

\item{...}{Not used.}
}
\value{
\code{x}, updated with a new or removed frosting postprocessor
}
\description{
Add frosting to a workflow
}
\examples{
# Try out frosting validators / constructors
 wf <- epi_workflow()
epipredict:::is_frosting(epipredict:::new_frosting()) # should be TRUE

 ## Three ways to add a frosting postprocessor
 wf1 <- epi_workflow(postprocessor = epipredict:::new_frosting())
 epipredict:::has_postprocessor(wf1) # should be TRUE
 epipredict:::has_postprocessor_frosting(wf1) # should be TRUE

 wf2 <- wf \%>\% epipredict:::add_postprocessor(epipredict:::new_frosting())
 epipredict:::has_postprocessor(wf2) # should be TRUE
 epipredict:::has_postprocessor_frosting(wf2) # should be TRUE

wf3 <- wf \%>\% add_frosting(epipredict:::new_frosting())
 epipredict:::has_postprocessor(wf3) # should be TRUE
 epipredict:::has_postprocessor_frosting(wf3) # should be TRUE

# Frosting can be created/added/removed
 f <- frosting()
 wf <- epi_workflow() \%>\% add_frosting(f)
epipredict:::has_postprocessor(wf) # should be TRUE
epipredict:::has_postprocessor_frosting(wf) # should be TRUE
 wf <- wf \%>\% remove_frosting()
 epipredict:::has_postprocessor(wf) # should be FALSE
 epipredict:::has_postprocessor_frosting(wf) # should be FALSE

}
