% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_classifier.R
\name{arx_class_args_list}
\alias{arx_class_args_list}
\title{ARX classifier argument constructor}
\usage{
arx_class_args_list(
  lags = c(0L, 7L, 14L),
  ahead = 7L,
  min_train_window = 20L,
  forecast_date = NULL,
  target_date = NULL,
  breaks = 0.25,
  horizon = 7L,
  method = c("rel_change", "linear_reg", "smooth_spline", "trend_filter"),
  log_scale = FALSE,
  additional_gr_args = list()
)
}
\arguments{
\item{lags}{Vector or List. Positive integers enumerating lags to use
in autoregressive-type models (in days).}

\item{ahead}{Integer. Number of time steps ahead (in days) of the forecast
date for which forecasts should be produced.}

\item{min_train_window}{Integer. The minimal amount of training
data (in the time unit of the \code{epi_df}) needed to produce a forecast.
If smaller, the forecaster will return \code{NA} predictions.}

\item{forecast_date}{Date. The date on which the forecast is created.
The default \code{NULL} will attempt to determine this automatically.}

\item{target_date}{Date. The date for which the forecast is intended.
The default \code{NULL} will attempt to determine this automatically.}

\item{breaks}{Vector. A vector of breaks to turn real-valued growth rates
into discrete classes. The default gives binary upswing classification
as in \href{https://doi.org/10.1073/pnas.2111453118}{McDonald, Bien, Green, Hu, et al.}.
This coincides with the default \code{trainer = parsnip::logistic_reg()}
argument in \code{\link[=arx_classifier]{arx_classifier()}}. However, multiclass classification is also
supported (e.g. with \code{breaks = c(-.2, .25)}) provided that
\code{trainer = parsnip::multinom_reg()} is used as well.}

\item{horizon}{Scalar integer. This is passed to the \code{h} argument of
\code{\link[epiprocess:growth_rate]{epiprocess::growth_rate()}}. It determines the amount of data used to
calculate the growth rate.}

\item{method}{Character. Options available for growth rate calculation.}

\item{log_scale}{Scalar logical. Whether to compute growth rates on the
log scale.}

\item{additional_gr_args}{List. Optional arguments controlling growth rate
calculation. See \code{\link[epiprocess:growth_rate]{epiprocess::growth_rate()}} and the related Vignette for
more details.}
}
\value{
A list containing updated parameter choices with class \code{arx_alist}.
}
\description{
Constructs a list of arguments for \code{\link[=arx_classifier]{arx_classifier()}}.
}
\examples{
arx_args_list()
arx_args_list(symmetrize = FALSE)
arx_args_list(levels = c(.1, .3, .7, .9), min_train_window = 120)
}
