% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_classifier.R
\name{arx_classifier}
\alias{arx_classifier}
\title{Direct autoregressive classifier with covariates}
\usage{
arx_classifier(
  epi_data,
  outcome,
  predictors,
  trainer = parsnip::logistic_reg(),
  args_list = arx_class_args_list()
)
}
\arguments{
\item{epi_data}{An \code{epi_df} object}

\item{outcome}{A character (scalar) specifying the outcome (in the
\code{epi_df}).}

\item{predictors}{A character vector giving column(s) of predictor
variables.}

\item{trainer}{A \code{{parsnip}} model describing the type of estimation.
For now, we enforce \code{mode = "classification"}. Typical values are
\code{\link[parsnip:logistic_reg]{parsnip::logistic_reg()}} or \code{\link[parsnip:multinom_reg]{parsnip::multinom_reg()}}. More complicated
trainers like \code{\link[parsnip:naive_Bayes]{parsnip::naive_Bayes()}} or \code{\link[parsnip:rand_forest]{parsnip::rand_forest()}} can
also be used.}

\item{args_list}{A list of customization arguments to determine
the type of forecasting model. See \code{\link[=arx_args_list]{arx_args_list()}}.}
}
\value{
A list with (1) \code{predictions} an \code{epi_df} of predicted classes
and (2) \code{epi_workflow}, a list that encapsulates the entire estimation
workflow
}
\description{
This is an autoregressive classification model for
\link[epiprocess:epi_df]{epiprocess::epi_df} data. It does "direct" forecasting, meaning
that it estimates a class at a particular target horizon.
}
\examples{
jhu <- case_death_rate_subset \%>\%
  dplyr::filter(time_value >= as.Date("2021-12-01"))

out <- arx_classifier(jhu, "death_rate", c("case_rate", "death_rate"))

out <- arx_classifier(
  jhu,
  "death_rate",
  c("case_rate", "death_rate"),
  trainer = parsnip::multinom_reg(),
  args_list = arx_class_args_list(
    breaks = c(-.2, .25), ahead = 14,
    horizon = 14, method = "linear_reg"
  )
)
}
