% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_forecaster.R
\name{arx_fcast_epi_workflow}
\alias{arx_fcast_epi_workflow}
\title{Create a template \code{arx_forecaster} workflow}
\usage{
arx_fcast_epi_workflow(
  epi_data,
  outcome,
  predictors = outcome,
  trainer = linear_reg(),
  args_list = arx_args_list()
)
}
\arguments{
\item{epi_data}{An \code{epi_df} object}

\item{outcome}{A character (scalar) specifying the outcome (in the
\code{epi_df}).}

\item{predictors}{A character vector giving column(s) of predictor variables.
This defaults to the \code{outcome}. However, if manually specified, only those variables
specifically mentioned will be used. (The \code{outcome} will not be added.)
By default, equals the outcome. If manually specified, does not add the
outcome variable, so make sure to specify it.}

\item{trainer}{A \code{{parsnip}} model describing the type of estimation. For
now, we enforce \code{mode = "regression"}. May be \code{NULL} if you'd like to
decide later.}

\item{args_list}{A list of customization arguments to determine
the type of forecasting model. See \code{\link[=arx_args_list]{arx_args_list()}}.}
}
\value{
An unfitted \code{epi_workflow}.
}
\description{
This function creates an unfit workflow for use with \code{\link[=arx_forecaster]{arx_forecaster()}}.
It is useful if you want to make small modifications to that forecaster
before fitting and predicting. Supplying a trainer to the function
may alter the returned \code{epi_workflow} object (e.g., if you intend to
use \code{\link[=quantile_reg]{quantile_reg()}}) but can be omitted.
}
\examples{
library(dplyr)
jhu <- covid_case_death_rates \%>\%
  filter(time_value >= as.Date("2021-12-01"))

arx_fcast_epi_workflow(
  jhu, "death_rate",
  c("case_rate", "death_rate")
)

arx_fcast_epi_workflow(jhu, "death_rate",
  c("case_rate", "death_rate"),
  trainer = quantile_reg(),
  args_list = arx_args_list(quantile_levels = 1:9 / 10)
)
}
\seealso{
\code{\link[=arx_forecaster]{arx_forecaster()}}, \code{\link[=arx_args_list]{arx_args_list()}}
}
