% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_workflow.R
\name{augment.epi_workflow}
\alias{augment.epi_workflow}
\title{Augment data with predictions}
\usage{
\method{augment}{epi_workflow}(x, new_data, ...)
}
\arguments{
\item{x}{A trained epi_workflow}

\item{new_data}{A epi_df of predictors}

\item{...}{Arguments passed on to the predict method.}
}
\value{
new_data with additional columns containing the predicted values
}
\description{
Augment data with predictions
}
