% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot-epipred}
\alias{autoplot-epipred}
\alias{autoplot.epi_workflow}
\alias{autoplot.canned_epipred}
\alias{plot.epi_workflow}
\alias{plot.canned_epipred}
\title{Automatically plot an \code{epi_workflow} or \code{canned_epipred} object}
\usage{
\method{autoplot}{epi_workflow}(
  object,
  predictions = NULL,
  observed_response = NULL,
  .levels = c(0.5, 0.8, 0.9),
  ...,
  .color_by = c("all_keys", "geo_value", "other_keys", ".response", "all", "none"),
  .facet_by = c(".response", "other_keys", "all_keys", "geo_value", "all", "none"),
  .base_color = "dodgerblue4",
  .point_pred_color = "orange",
  .facet_filter = NULL,
  .max_facets = deprecated()
)

\method{autoplot}{canned_epipred}(
  object,
  observed_response = NULL,
  ...,
  .color_by = c("all_keys", "geo_value", "other_keys", ".response", "all", "none"),
  .facet_by = c(".response", "other_keys", "all_keys", "geo_value", "all", "none"),
  .base_color = "dodgerblue4",
  .point_pred_color = "orange",
  .facet_filter = NULL,
  .max_facets = deprecated()
)

\method{plot}{epi_workflow}(x, ...)

\method{plot}{canned_epipred}(x, ...)
}
\arguments{
\item{object, x}{An \code{epi_workflow}}

\item{predictions}{A data frame with predictions. If \code{NULL}, only the
original data is shown.}

\item{observed_response}{An epi_df of the data to plot against. This is for the case
where you have the actual results to compare the forecast against.}

\item{.levels}{A numeric vector of levels to plot for any prediction bands.
More than 3 levels begins to be difficult to see.}

\item{...}{Ignored}

\item{.color_by}{Which variables should determine the color(s) used to plot
lines. Options include:
\itemize{
\item \code{all_keys} - the default uses the interaction of any key variables
including the \code{geo_value}
\item \code{geo_value} - \code{geo_value} only
\item \code{other_keys} - any available keys that are not \code{geo_value}
\item \code{.response} - the numeric variables (same as the y-axis)
\item \code{all} - uses the interaction of all keys and numeric variables
\item \code{none} - no coloring aesthetic is applied
}}

\item{.facet_by}{Similar to \code{.color_by} except that the default is to
display the response.}

\item{.base_color}{If available, prediction bands will be shown with this
color.}

\item{.point_pred_color}{If available, point forecasts will be shown with
this color.}

\item{.facet_filter}{Select which facets will be displayed. Especially
useful for when there are many \code{geo_value}'s or keys. This is a
<\code{\link[=args_data_masking]{rlang}}> expression along the lines of \code{\link[dplyr:filter]{dplyr::filter()}}.
However, it must be a single expression combined with the \code{&} operator. This
contrasts to the typical use case which allows multiple comma-separated expressions
which are implicitly combined with \code{&}. When multiple variables are selected
with \code{...}, their names can be filtered in combination with other factors
by using \code{.response_name}. See the examples below.}

\item{.max_facets}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\description{
For a fit workflow, the training data will be displayed, the response by
default. If \code{predictions} is not \code{NULL} then point and interval forecasts
will be shown as well. Unfit workflows will result in an error, (you
can simply call \code{autoplot()} on the original \code{epi_df}).
}
\examples{
jhu <- covid_case_death_rates \%>\%
  filter(time_value >= as.Date("2021-11-01"))

r <- epi_recipe(jhu) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_lag(case_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_naomit()

f <- frosting() \%>\%
  layer_residual_quantiles() \%>\%
  layer_threshold(starts_with(".pred")) \%>\%
  layer_add_target_date()

wf <- epi_workflow(r, linear_reg(), f) \%>\% fit(jhu)

autoplot(wf)

latest <- jhu \%>\% filter(time_value >= max(time_value) - 14)
preds <- predict(wf, latest)
autoplot(wf, preds, .facet_filter = geo_value \%in\% c("ca", "ny", "de", "mt"))

# ------- Show multiple horizons

p <- lapply(c(7, 14, 21, 28), function(h) {
  r <- epi_recipe(jhu) \%>\%
    step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
    step_epi_ahead(death_rate, ahead = h) \%>\%
    step_epi_lag(case_rate, lag = c(0, 7, 14)) \%>\%
    step_epi_naomit()
  ewf <- epi_workflow(r, linear_reg(), f) \%>\% fit(jhu)
  forecast(ewf)
})

p <- do.call(rbind, p)
autoplot(wf, p, .facet_filter = geo_value \%in\% c("ca", "ny", "de", "mt"))

# ------- Plotting canned forecaster output

jhu <- covid_case_death_rates \%>\%
  filter(time_value >= as.Date("2021-11-01"))
flat <- flatline_forecaster(jhu, "death_rate")
autoplot(flat, .facet_filter = geo_value \%in\% c("ca", "ny", "de", "mt"))

arx <- arx_forecaster(jhu, "death_rate", c("case_rate", "death_rate"),
  args_list = arx_args_list(ahead = 14L)
)
autoplot(arx, .facet_filter = geo_value \%in\% c("ca", "ny", "de", "mt", "mo", "in"))
}
