% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatological_forecaster.R
\name{climate_args_list}
\alias{climate_args_list}
\title{Climatological forecaster argument constructor}
\usage{
climate_args_list(
  forecast_date = NULL,
  forecast_horizon = 0:4,
  time_type = c("epiweek", "week", "month", "day"),
  center_method = c("median", "mean"),
  window_size = 3L,
  quantile_levels = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95),
  symmetrize = FALSE,
  nonneg = TRUE,
  quantile_by_key = character(0L),
  ...
)
}
\arguments{
\item{forecast_date}{Date. The date from which the forecast is occurring.
The default \code{NULL} will determine this automatically from either
\enumerate{
\item the maximum time value for which there's data if there is no latency
adjustment (the default case), or
\item the \code{as_of} date of \code{epi_data} if \code{adjust_latency} is
non-\code{NULL}.
}}

\item{forecast_horizon}{Vector of integers giving the number of time steps,
in units of the \code{time_type},
from the \code{reference_date} for which predictions should be produced.}

\item{time_type}{The duration over which time aggregation should be performed.}

\item{center_method}{The measure of center to be calculated over the time
window.}

\item{window_size}{Scalar integer. How many time units on each side should
be included. For example, if \code{window_size = 3} and \code{time_type = "day"},
then on each day in the data, the center will be calculated using 3 days
before and three days after. So, in this case, it operates like a weekly
rolling average, centered at each day.}

\item{quantile_levels}{Vector or \code{NULL}. A vector of probabilities to produce
prediction intervals. These are created by computing the quantiles of
training residuals. A \code{NULL} value will result in point forecasts only.}

\item{symmetrize}{Logical. The default \code{TRUE} calculates
symmetric prediction intervals. This argument only applies when
residual quantiles are used. It is not applicable with
\code{trainer = quantile_reg()}, for example.}

\item{nonneg}{Logical. The default \code{TRUE} enforces nonnegative predictions
by hard-thresholding at 0.}

\item{quantile_by_key}{Character vector. Groups residuals by listed keys
before calculating residual quantiles. See the \code{by_key} argument to
\code{\link[=layer_residual_quantiles]{layer_residual_quantiles()}} for more information. The default,
\code{character(0)} performs no grouping. This argument only applies when
residual quantiles are used. It is not applicable with
\code{trainer = quantile_reg()}, for example.}

\item{...}{Further arguments passed to or from other methods (not currently
used).}
}
\value{
A list containing updated parameter choices with class \code{climate_alist}.
}
\description{
Climatological forecaster argument constructor
}
\examples{
climate_args_list()
climate_args_list(
  forecast_horizon = 0:10,
  quantile_levels = c(.01, .025, 1:19 / 20, .975, .99)
)

}
\seealso{
\code{\link[=climatological_forecaster]{climatological_forecaster()}}, \code{\link[=step_climate]{step_climate()}}
}
