% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{count_single_column}
\alias{count_single_column}
\title{get the location of the last real value}
\usage{
count_single_column(col)
}
\arguments{
\item{col}{the relevant column}
}
\description{
get the location of the last real value
}
\keyword{internal}
