% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_leads_and_lags.R
\name{create_lags_and_leads}
\alias{create_lags_and_leads}
\title{Create lags and leads of predictors and response}
\usage{
create_lags_and_leads(x, y, xy_lags, y_leads)
}
\arguments{
\item{x}{Data frame or matrix. Predictor variables.}

\item{y}{Response vector. Typical usage will "lead" y by the
number of steps forward for the prediction horizon (ahead).}

\item{xy_lags}{Vector or list. If a vector, the lags will apply
to each column of \code{x} and to \code{y}. If a list, it must be of length
\code{ncol(x)+1} and each component will apply to the requisite predictor.
A \code{NULL} list element will remove that variable completely from the
result. Negative values will "lead" the variable.}

\item{y_leads}{Scalar or vector. If a scalar, we "lead" \code{y} by this
amount. A vector will produce multiple columns of \code{y} if this is
useful for your model. Negative values will "lag" the variable.}
}
\value{
A \code{data.frame}.
}
\description{
Create lags and leads of predictors and response
}
\examples{

x <- 1:20
y <- -20:-1
create_lags_and_leads(x, y, c(1, 2), 1)
create_lags_and_leads(x, y, list(c(1, 2), 1), 1)
create_lags_and_leads(x, y, list(c(-1, 1), NULL), 1)
create_lags_and_leads(x, y, c(1, 2), c(0, 1))
}
