% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-layer.R
\name{create_layer}
\alias{create_layer}
\title{Create a new layer}
\usage{
create_layer(name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{name}{Either a name without extension, or \code{NULL} to create the
paired file based on currently open file in the script editor. If
the \verb{R/} file is open, \code{use_test()} will create/open the corresponding
test file; if the test file is open, \code{use_r()} will create/open the
corresponding \verb{R/} file.}

\item{open}{Whether to open the file for interactive editing.}
}
\description{
This function creates the skeleton for a new \code{frosting} layer. When called
inside a package, it will create an R script in the \verb{R/} directory,
fill in the name of the layer, and open the file.
}
\examples{
\dontrun{

  # Note: running this will write `layer_strawberry.R` to
  # the `R/` directory of your current project
  create_layer("strawberry")
}

}
