% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_quantiles.R
\name{dist_quantiles}
\alias{dist_quantiles}
\title{A distribution parameterized by a set of quantiles}
\usage{
dist_quantiles(x, tau)
}
\arguments{
\item{x}{A vector of values}

\item{tau}{A vector of probabilities corresponding to \code{x}}
}
\description{
A distribution parameterized by a set of quantiles
}
\examples{
dstn <- dist_quantiles(list(1:4, 8:11), list(c(.2, .4, .6, .8)))
quantile(dstn, p = c(.1, .25, .5, .9))
median(dstn)

# it's a bit annoying to inspect the data
vctrs::vec_data(vctrs::vec_data(dstn[1])[[1]])
}
