% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_quantiles.R
\name{extrapolate_quantiles}
\alias{extrapolate_quantiles}
\title{Summarize a distribution with a set of quantiles}
\usage{
extrapolate_quantiles(x, probs, replace_na = TRUE, ...)
}
\arguments{
\item{x}{a \code{distribution} vector}

\item{probs}{a vector of probabilities at which to calculate quantiles}

\item{replace_na}{logical. If \code{x} contains \code{NA}'s, these are imputed if
possible (if \code{TRUE}) or retained (if \code{FALSE}). This only effects
elements of class \code{dist_quantiles}.}

\item{...}{additional arguments passed on to the \code{quantile} method}
}
\value{
a \code{distribution} vector containing \code{dist_quantiles}. Any elements
of \code{x} which were originally \code{dist_quantiles} will now have a superset
of the original \code{quantile_values} (the union of those and \code{probs}).
}
\description{
Summarize a distribution with a set of quantiles
}
\examples{
library(distributional)
dstn <- dist_normal(c(10, 2), c(5, 10))
extrapolate_quantiles(dstn, probs = c(.25, 0.5, .75))

dstn <- dist_quantiles(list(1:4, 8:11), list(c(.2, .4, .6, .8)))
# because this distribution is already quantiles, any extra quantiles are
# appended
extrapolate_quantiles(dstn, probs = c(.25, 0.5, .75))

dstn <- c(
  dist_normal(c(10, 2), c(5, 10)),
  dist_quantiles(list(1:4, 8:11), list(c(.2, .4, .6, .8)))
)
extrapolate_quantiles(dstn, probs = c(.25, 0.5, .75))
}
