% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_quantiles.R
\name{extrapolate_quantiles}
\alias{extrapolate_quantiles}
\title{Summarize a distribution with a set of quantiles}
\usage{
extrapolate_quantiles(x, probs, replace_na = TRUE, ...)
}
\arguments{
\item{x}{A vector of class \code{quantile_pred}.}

\item{probs}{a vector of probabilities at which to calculate quantiles}

\item{replace_na}{logical. If \code{x} contains \code{NA}'s, these are imputed if
possible (if \code{TRUE}) or retained (if \code{FALSE}).}

\item{...}{additional arguments passed on to the \code{quantile} method}
}
\value{
a \code{quantile_pred} vector. Each element
of \code{x} will now have a superset
of the original \code{quantile_values} (the union of those and \code{probs}).
}
\description{
This function takes a \code{quantile_pred} vector and returns the same
type of object, expanded to include
\emph{additional} quantiles computed at \code{probs}. If you want behaviour more
similar to \code{\link[stats:quantile]{stats::quantile()}}, then \code{quantile(x,...)} may be more
appropriate.
}
\examples{
dstn <- quantile_pred(rbind(1:4, 8:11), c(.2, .4, .6, .8))
# extra quantiles are appended
as_tibble(extrapolate_quantiles(dstn, probs = c(.25, 0.5, .75)))
}
