% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatline_epi_forecaster.R
\name{flatline_args_list}
\alias{flatline_args_list}
\title{Flatline forecaster argument constructor}
\usage{
flatline_args_list(
  ahead = 7L,
  min_train_window = 20L,
  forecast_date = NULL,
  target_date = NULL,
  levels = c(0.05, 0.95),
  symmetrize = TRUE,
  nonneg = TRUE
)
}
\arguments{
\item{ahead}{Integer. Number of time steps ahead of the forecast date
for which forecasts should be produced.}

\item{min_train_window}{Integer. The minimal amount of training
data needed to produce a forecast. If smaller, the forecaster will return
\code{NA} predictions.}

\item{forecast_date}{The date on which the forecast is created. The default
\code{NULL} will attempt to determine this automatically.}

\item{target_date}{The date for which the forecast is intended. The default
\code{NULL} will attempt to determine this automatically.}

\item{levels}{Vector or \code{NULL}. A vector of probabilities to produce
prediction intervals. These are created by computing the quantiles of
training residuals. A \code{NULL} value will result in point forecasts only.}

\item{symmetrize}{Logical. The default \code{TRUE} calculates
symmetric prediction intervals.}

\item{nonneg}{Logical. The default \code{TRUE} enforeces nonnegative predictions
by hard-thresholding at 0.}
}
\value{
A list containing updated parameter choices.
}
\description{
Constructs a list of arguments for \code{\link[=flatline_epi_forecaster]{flatline_epi_forecaster()}}.
}
\examples{
flatline_args_list()
flatline_args_list(symmetrize = FALSE)
flatline_args_list(levels = c(.1, .3, .7, .9), min_train_window = 120)
}
