% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatline_forecaster.R
\name{flatline_args_list}
\alias{flatline_args_list}
\title{Flatline forecaster argument constructor}
\usage{
flatline_args_list(
  ahead = 7L,
  n_training = Inf,
  forecast_date = NULL,
  target_date = NULL,
  quantile_levels = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95),
  symmetrize = TRUE,
  nonneg = TRUE,
  quantile_by_key = character(0L),
  ...
)
}
\arguments{
\item{ahead}{Integer. Unlike \code{\link[=arx_forecaster]{arx_forecaster()}}, this doesn't have any effect
on the predicted values. Predictions are always the most recent observation.
However, this \emph{does} impact the residuals stored in the object. Residuals
are calculated based on this number to mimic how badly you would have done.
So for example, \code{ahead = 7} will create residuals by comparing values
7 days apart.}

\item{n_training}{Integer. An upper limit for the number of rows per
key that are used for training
(in the time unit of the \code{epi_df}).}

\item{target_date}{Date. The date that is being forecast. The default \code{NULL}
will determine this automatically as \code{forecast_date + ahead}.}

\item{quantile_levels}{Vector or \code{NULL}. A vector of probabilities to produce
prediction intervals. These are created by computing the quantiles of
training residuals. A \code{NULL} value will result in point forecasts only.}

\item{symmetrize}{Logical. The default \code{TRUE} calculates
symmetric prediction intervals. This argument only applies when
residual quantiles are used. It is not applicable with
\code{trainer = quantile_reg()}, for example.}

\item{nonneg}{Logical. The default \code{TRUE} enforces nonnegative predictions
by hard-thresholding at 0.}

\item{quantile_by_key}{Character vector. Groups residuals by listed keys
before calculating residual quantiles. See the \code{by_key} argument to
\code{\link[=layer_residual_quantiles]{layer_residual_quantiles()}} for more information. The default,
\code{character(0)} performs no grouping. This argument only applies when
residual quantiles are used. It is not applicable with
\code{trainer = quantile_reg()}, for example.}

\item{...}{Space to handle future expansions (unused).}
}
\value{
A list containing updated parameter choices with class \code{flatline_alist}.
}
\description{
Constructs a list of arguments for \code{\link[=flatline_forecaster]{flatline_forecaster()}}.
}
\examples{
flatline_args_list()
flatline_args_list(symmetrize = FALSE)
flatline_args_list(quantile_levels = c(.1, .3, .7, .9), n_training = 120)
}
