% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatline_forecaster.R
\name{flatline_args_list}
\alias{flatline_args_list}
\title{Flatline forecaster argument constructor}
\usage{
flatline_args_list(
  ahead = 7L,
  n_training = Inf,
  forecast_date = NULL,
  target_date = NULL,
  quantile_levels = c(0.05, 0.95),
  symmetrize = TRUE,
  nonneg = TRUE,
  quantile_by_key = character(0L),
  nafill_buffer = Inf,
  ...
)
}
\arguments{
\item{ahead}{Integer. Unlike \code{\link[=arx_forecaster]{arx_forecaster()}}, this doesn't have any effect
on the predicted values. Predictions are always the most recent observation.
However, this \emph{does} impact the residuals stored in the object. Residuals
are calculated based on this number to mimic how badly you would have done.
So for example, \code{ahead = 7} will create residuals by comparing values
7 days apart.}

\item{n_training}{Integer. An upper limit for the number of rows per
key that are used for training
(in the time unit of the \code{epi_df}).}

\item{forecast_date}{Date. The date on which the forecast is created.  The
default \code{NULL} will attempt to determine this automatically either as the
max time value if there is no latency adjustment, or as the \code{as_of} of
\code{epi_data} if \code{adjust_latency} is non-\code{NULL}.}

\item{target_date}{Date. The date for which the forecast is intended. The
default \code{NULL} will attempt to determine this automatically as
\code{forecast_date + ahead}.}

\item{quantile_levels}{Vector or \code{NULL}. A vector of probabilities to produce
prediction intervals. These are created by computing the quantiles of
training residuals. A \code{NULL} value will result in point forecasts only.}

\item{symmetrize}{Logical. The default \code{TRUE} calculates
symmetric prediction intervals. This argument only applies when
residual quantiles are used. It is not applicable with
\code{trainer = quantile_reg()}, for example.}

\item{nonneg}{Logical. The default \code{TRUE} enforces nonnegative predictions
by hard-thresholding at 0.}

\item{quantile_by_key}{Character vector. Groups residuals by listed keys
before calculating residual quantiles. See the \code{by_key} argument to
\code{\link[=layer_residual_quantiles]{layer_residual_quantiles()}} for more information. The default,
\code{character(0)} performs no grouping. This argument only applies when
residual quantiles are used. It is not applicable with
\code{trainer = quantile_reg()}, for example.}

\item{nafill_buffer}{At predict time, recent values of the training data
are used to create a forecast. However, these can be \code{NA} due to, e.g.,
data latency issues. By default, any missing values will get filled with
less recent data. Setting this value to \code{NULL} will result in 1 extra
recent row (beyond those required for lag creation) to be used. Note that
we require at least \code{min(lags)} rows of recent data per \code{geo_value} to
create a prediction. For this reason, setting \code{nafill_buffer < min(lags)}
will be treated as \emph{additional} allowed recent data rather than the
total amount of recent data to examine.}

\item{...}{Space to handle future expansions (unused).}
}
\value{
A list containing updated parameter choices with class \code{flatline_alist}.
}
\description{
Constructs a list of arguments for \code{\link[=flatline_forecaster]{flatline_forecaster()}}.
}
\examples{
flatline_args_list()
flatline_args_list(symmetrize = FALSE)
flatline_args_list(quantile_levels = c(.1, .3, .7, .9), n_training = 120)
}
