% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatline_forecaster.R
\name{flatline_forecaster}
\alias{flatline_forecaster}
\title{Predict the future with today's value}
\usage{
flatline_forecaster(
  epi_data,
  outcome,
  args_list = flatline_args_list(),
  .verbose = TRUE
)
}
\arguments{
\item{epi_data}{An \link[epiprocess:epi_df]{epiprocess::epi_df}}

\item{outcome}{A scalar character for the column name we wish to predict.}

\item{args_list}{A list of dditional arguments as created by the
\code{\link[=flatline_args_list]{flatline_args_list()}} constructor function.}

\item{.verbose}{If true, prints the arguments used in the workflow.}
}
\value{
A data frame of point (and optionally interval) forecasts at a single
ahead (unique horizon) for each unique combination of \code{key_vars}.
}
\description{
This is a simple forecasting model for
\link[epiprocess:epi_df]{epiprocess::epi_df} data. It uses the most recent observation as the
forcast for any future date, and produces intervals based on the quantiles
of the residuals of such a "flatline" forecast over all available training
data.
}
\details{
By default, the predictive intervals are computed separately for each
combination of key values (\code{geo_value} + any additional keys) in the
\code{epi_data} argument.

This forecaster is very similar to that used by the
\href{https://covid19forecasthub.org}{COVID19ForecastHub}
}
\examples{
jhu <- case_death_rate_subset \%>\%
  dplyr::filter(time_value >= as.Date("2021-12-01"))

out <- flatline_forecaster(jhu, "death_rate")
}
