% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_workflow.R
\name{forecast.epi_workflow}
\alias{forecast.epi_workflow}
\title{Produce a forecast from just an epi workflow}
\usage{
\method{forecast}{epi_workflow}(object)
}
\arguments{
\item{object}{An epi workflow.}
}
\value{
A forecast tibble.
}
\description{
\code{forecast.epi_workflow} predicts by restricting the training data to the
latest available data, and predicting on that. It binds together
\code{get_test_data()} and \code{predict()}.
}
\examples{
jhu <- covid_case_death_rates \%>\%
  filter(time_value > "2021-08-01")

r <- epi_recipe(jhu) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_naomit()

epi_workflow(r, parsnip::linear_reg()) \%>\%
  fit(jhu) \%>\%
  forecast()
}
