% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frosting.R
\name{frosting}
\alias{frosting}
\title{Create frosting for postprocessing predictions}
\usage{
frosting(layers = NULL, requirements = NULL)
}
\arguments{
\item{layers}{Must be \code{NULL}.}

\item{requirements}{Must be \code{NULL}.}
}
\value{
A frosting object.
}
\description{
This generates a postprocessing container (much like \code{recipes::recipe()})
to hold steps for postprocessing predictions.
}
\details{
The arguments are currently placeholders and must be NULL
}
\examples{
library(dplyr)
# Toy example to show that frosting can be created and added for postprocessing
f <- frosting()
wf <- epi_workflow() \%>\% add_frosting(f)

# A more realistic example
jhu <- case_death_rate_subset \%>\%
  filter(time_value > "2021-11-01", geo_value \%in\% c("ak", "ca", "ny"))

r <- epi_recipe(jhu) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_naomit()

wf <- epi_workflow(r, parsnip::linear_reg()) \%>\% fit(jhu)

f <- frosting() \%>\%
  layer_predict() \%>\%
  layer_naomit(.pred)

wf1 <- wf \%>\% add_frosting(f)

p <- forecast(wf1)
p
}
