% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_forecast_date_in_layer}
\alias{get_forecast_date_in_layer}
\title{get the target date while in a layer}
\usage{
get_forecast_date_in_layer(this_recipe, workflow_max_time_value, new_data)
}
\arguments{
\item{this_recipe}{the recipe to check for \code{step_adjust_latency}}

\item{workflow_max_time_value}{the \code{max_time} value coming out of the fit
workflow (this will be the maximal time value in a potentially different
dataset)}

\item{new_data}{the data we're currently working with, from which we'll take
a potentially different max_time_value}
}
\description{
get the target date while in a layer
}
\keyword{internal}
