% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latency.R
\name{get_latent_column_tibble}
\alias{get_latent_column_tibble}
\title{find the columns added with the lags or aheads, and the amounts they have
been changed}
\usage{
get_latent_column_tibble(
  shift_cols,
  new_data,
  forecast_date,
  latency,
  sign_shift,
  info,
  epi_keys_checked,
  call = caller_env()
)
}
\arguments{
\item{shift_cols}{a list of columns to operate on, as created by \code{construct_shift_tibble}}

\item{new_data}{the data transformed so far}

\item{forecast_date}{the forecast date}

\item{latency}{\code{NULL}, int, or vector, as described in \code{step_epi_latency}}

\item{sign_shift}{-1 if ahead, 1 if lag}
}
\value{
a tibble with columns \code{column} (relevant shifted names), \code{shift} (the
amount that one is shifted), \code{latency} (original columns difference between
the max \code{time_value} and \code{forecast_date} (on a per-initial column basis)),
\code{effective_shift} (shift+latency), and \code{new_name} (adjusted names with the
effective_shift)
}
\description{
find the columns added with the lags or aheads, and the amounts they have
been changed
}
\keyword{internal}
