% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Get test data for prediction based on longest lag period}
\usage{
get_test_data(recipe, x)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{x}{A data frame, tibble, or epi_df data set.}
}
\value{
A tibble with columns \code{geo_value}, \code{time_value}
and other variables in the original dataset.
}
\description{
Based on the longest lag period in the recipe,
\code{get_test_data()} creates a tibble in \link[epiprocess:epi_df]{epiprocess::epi_df}
format with columns \code{geo_value}, \code{time_value}
and other variables in the original dataset,
which will be used to create test data.
}
\examples{
# create recipe
 rec <- epi_recipe(case_death_rate_subset) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_lag(case_rate, lag = c(0, 7, 14))
 get_test_data(recipe = rec, x = case_death_rate_subset)
}
