% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Get test data for prediction based on longest lag period}
\usage{
get_test_data(
  recipe,
  x,
  fill_locf = FALSE,
  n_recent = NULL,
  forecast_date = max(x$time_value)
)
}
\arguments{
\item{recipe}{A recipe object.}

\item{x}{An epi_df. The typical usage is to
pass the same data as that used for fitting the recipe.}

\item{fill_locf}{Logical. Should we use \code{locf} to fill in missing data?}

\item{n_recent}{Integer or NULL. If filling missing data with \code{locf = TRUE},
how far back are we willing to tolerate missing data? Larger values allow
more filling. The default \code{NULL} will determine this from the
the \code{recipe}. For example, suppose \code{n_recent = 3}, then if the
3 most recent observations in any \code{geo_value} are all \code{NA}’s, we won’t be
able to fill anything, and an error message will be thrown. (See details.)}

\item{forecast_date}{By default, this is set to the maximum
\code{time_value} in \code{x}. But if there is data latency such that recent \code{NA}'s
should be filled, this may be \emph{after} the last available \code{time_value}.}
}
\value{
An object of the same type as \code{x} with columns \code{geo_value}, \code{time_value}, any additional
keys, as well other variables in the original dataset.
}
\description{
Based on the longest lag period in the recipe,
\code{get_test_data()} creates an \link[epiprocess:epi_df]{epi_df}
with columns \code{geo_value}, \code{time_value}
and other variables in the original dataset,
which will be used to create features necessary to produce forecasts.
}
\details{
The minimum required (recent) data to produce a forecast is equal to
the maximum lag requested (on any predictor) plus the longest horizon
used if growth rate calculations are requested by the recipe. This is
calculated internally.

It also optionally fills missing values
using the last-observation-carried-forward (LOCF) method. If this
is not possible (say because there would be only \code{NA}'s in some location),
it will produce an error suggesting alternative options to handle missing
values with more advanced techniques.
}
\examples{
# create recipe
rec <- epi_recipe(covid_case_death_rates) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_lag(case_rate, lag = c(0, 7, 14))
get_test_data(recipe = rec, x = covid_case_death_rates)
}
