% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grad_employ_subset}
\alias{grad_employ_subset}
\title{Subset of Statistics Canada median employment income for postsecondary graduates}
\format{
A tibble with 1607 rows and 7 variables:
\describe{
\item{geo_value}{The province in Canada associated with each
row of measurements.}
\item{time_value}{The time value, a year integer in YYYY format}
\item{edu_qual}{The education qualification}
\item{age_group}{The age group; either 15 to 34 or 35 to 64}
\item{num_graduates}{The number of graduates for the given row of characteristics}
\item{med_income_2y}{The median employment income two years after graduation}
\item{med_income_5y}{The median employment income five years after graduation}
}
}
\source{
This object contains modified data from the following Statistics Canada
data table: \href{https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=3710011501}{
Characteristics and median employment income of longitudinal cohorts of postsecondary
graduates two and five years after graduation, by educational qualification and
field of study (primary groupings)
}

Modifications:
\itemize{
\item Only provincial-level geo_values are kept
\item Only age group, field of study, and educational qualification are kept as
covariates. For the remaining covariates, we keep aggregated values and
drop the level-specific rows.
\item No modifications were made to the time range of the data
}
}
\usage{
grad_employ_subset
}
\description{
Subset of Statistics Canada median employment income for postsecondary graduates
}
\keyword{datasets}
