% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_naomit.R
\name{layer_naomit}
\alias{layer_naomit}
\title{Omit \code{NA}s from predictions or other columns}
\usage{
layer_naomit(frosting, ..., id = rand_id("naomit"))
}
\arguments{
\item{frosting}{a \code{frosting} postprocessor}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables. Typical usage is \code{.pred} to remove
any rows with \code{NA} predictions.}

\item{id}{a random id string}
}
\value{
an updated \code{frosting} postprocessor
}
\description{
Omit \code{NA}s from predictions or other columns
}
\examples{
 jhu <- case_death_rate_subset \%>\%
  dplyr::filter(time_value > "2021-11-01", geo_value \%in\% c("ak", "ca", "ny"))

r <- epi_recipe(jhu) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14)) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  recipes::step_naomit(recipes::all_predictors()) \%>\%
  recipes::step_naomit(recipes::all_outcomes(), skip = TRUE)

wf <- epi_workflow(r, parsnip::linear_reg()) \%>\%
 parsnip::fit(jhu)

latest <- get_test_data(recipe = r, x = jhu)

f <- epipredict:::frosting() \%>\%
     layer_predict() \%>\%
     layer_naomit(.pred)

wf1 <- wf \%>\% epipredict:::add_frosting(f)

p <- predict(wf1, latest)
p
}
