% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_nonneg_preds.R
\name{layer_nonneg_preds}
\alias{layer_nonneg_preds}
\title{Lower and upper thresholds for prediction values}
\usage{
layer_nonneg_preds(
  frosting,
  pred_lower = 0,
  pred_upper = Inf,
  ...,
  id = rand_id("nonneg_preds")
)
}
\arguments{
\item{frosting}{a \code{frosting} postprocessor}

\item{pred_lower}{Lower threshold for the prediction values. That is, any
predictions greater than or equal to this lower bound are retained.
Default value is \code{0}.}

\item{pred_upper}{Upper threshold for the prediction values. That is, any
predictions less than or equal to this upper bound are retained.
Default value is \code{Inf}.}

\item{id}{a random id string}
}
\value{
an updated \code{frosting} postprocessor
}
\description{
This postprocessing step is used to retain only those
prediction values that are within the specified lower and upper thresholds.
}
