% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_population_scaling.R
\name{layer_population_scaling}
\alias{layer_population_scaling}
\title{Revert population scaled prediction}
\usage{
layer_population_scaling(
  frosting,
  ...,
  df,
  by = NULL,
  df_pop_col,
  create_new = TRUE,
  suffix = "_original",
  .flag = TRUE,
  id = rand_id("population_scaling")
)
}
\arguments{
\item{frosting}{a \code{frosting} postprocessor}

\item{...}{the column(s) in the \code{epi_df} to scale back.}

\item{df}{a data frame that contains the population data used for scaling.}

\item{by}{A character vector of variables to join by. Column names should be
in lower case.
To join by different variables on the \code{epi_df} and \code{df}, use a named vector.
For example, by = c("geo_value" = "states") will match \code{epi_df$geo_value}
to \code{df$states}. To join by multiple variables, use a vector with length > 1.
For example, by = c("geo_value" = "states", "county" = "county") will match
\code{epi_df$geo_value} to \code{df$states} and \code{epi_df$county} to \code{df$county}.}

\item{df_pop_col}{the name of the column in the data frame \code{df} that
contains the population data and will be used for scaling.}

\item{create_new}{TRUE to create a new column and keep the original column
in the \code{epi_df}.}

\item{suffix}{a character. The suffix added to the column name if
\code{crete_new = TRUE}. Default to "_original".}

\item{.flag}{a logical to determine if the layer is added. Passed on to
\code{add_layer()}. Default \code{TRUE}.}

\item{id}{a random id string}
}
\value{
an updated \code{frosting} postprocessor with population scaled variables
reverted back to the original variable.
}
\description{
Revert population scaled prediction
}
