% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_residual_quantile.R
\name{layer_residual_quantile}
\alias{layer_residual_quantile}
\title{Creates predictions based on residual quantiles}
\usage{
layer_residual_quantile(
  frosting,
  probs,
  symmetrize,
  id = rand_id("residual_quantile")
)
}
\arguments{
\item{frosting}{a \code{frosting} postprocessor}

\item{probs}{numeric vector of probabilities with values in \link{0,1} referring to the desired quantile.}

\item{symmetrize}{logical. If \code{TRUE} then interval will be symmetrical.}

\item{id}{a random id string}
}
\value{
an updated \code{frosting} postprocessor with additional columns of the residual quantiles added to the prediction
}
\description{
Creates predictions based on residual quantiles
}
\examples{
TO-DO
}
