% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_unnest.R
\name{layer_unnest}
\alias{layer_unnest}
\title{Unnest prediction list-cols}
\usage{
layer_unnest(frosting, ..., id = rand_id("unnest"))
}
\arguments{
\item{frosting}{a \code{frosting} postprocessor}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}

\item{id}{a random id string}
}
\value{
an updated \code{frosting} postprocessor
}
\description{
Unnest prediction list-cols
}
