% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_quantiles.R
\name{pivot_quantiles}
\alias{pivot_quantiles}
\title{Pivot columns containing \code{dist_quantile} wider}
\usage{
pivot_quantiles(.data, ...)
}
\arguments{
\item{.data}{A data frame, or a data frame extension such as a tibble or
epi_df.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables. Any selected columns should}
}
\value{
An object of the same class as \code{.data}
}
\description{
Any selected columns that contain \code{dist_quantiles} will be "widened" with
the "taus" (quantile) serving as names and the values in the data frame.
When pivoting multiple columns, the original column name will be used as
a prefix.
}
\examples{
d1 <- c(dist_quantiles(1:3, 1:3 / 4), dist_quantiles(2:4, 1:3 / 4))
d2 <- c(dist_quantiles(2:4, 2:4 / 5), dist_quantiles(3:5, 2:4 / 5))
tib <- tibble(g = c("a", "b"), d1 = d1, d2 = d2)

pivot_quantiles(tib, c("d1", "d2"))
pivot_quantiles(tib, tidyselect::starts_with("d"))
pivot_quantiles(tib, d2)
}
