% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{filter}
\alias{mutate}
\alias{rename}
\alias{select}
\alias{as_epi_df}
\alias{key_colnames}
\alias{fit}
\alias{forecast}
\alias{tidy}
\alias{prep}
\alias{bake}
\alias{rand_id}
\alias{as_tibble}
\alias{tibble}
\alias{pivot_longer}
\alias{pivot_wider}
\alias{unnest}
\alias{as_tibble}
\alias{tidy}
\alias{quantile_pred}
\alias{extract_quantile_levels}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}}

  \item{epiprocess}{\code{\link[epiprocess:epi_df]{as_epi_df}}, \code{\link[epiprocess]{key_colnames}}}

  \item{generics}{\code{\link[generics]{fit}}, \code{\link[generics]{forecast}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{hardhat}{\code{\link[hardhat:quantile_pred]{extract_quantile_levels}}, \code{\link[hardhat]{quantile_pred}}}

  \item{recipes}{\code{\link[recipes]{bake}}, \code{\link[recipes]{prep}}, \code{\link[recipes]{rand_id}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}

  \item{tidyr}{\code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{unnest}}}
}}
