% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_climate.R
\name{roll_modular_multivec}
\alias{roll_modular_multivec}
\title{group col by .idx values and sum windows around each .idx value}
\usage{
roll_modular_multivec(col, .idx, weights, aggr, window_size, modulus)
}
\arguments{
\item{col}{the list of values indexed by \code{.idx}}

\item{.idx}{the relevant periodic part of time value, e.g. the week number}

\item{weights}{how much to weigh each particular datapoint}

\item{aggr}{the aggregation function, probably Quantile, mean or median}

\item{window_size}{the number of .idx entries before and after to include in
the aggregation}

\item{modulus}{the maximum value of \code{.idx}}
}
\description{
group col by .idx values and sum windows around each .idx value
}
\keyword{internal}
