% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-epidatasets.R
\docType{data}
\name{state_census}
\alias{state_census}
\title{State population data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57 rows and 4 columns.
}
\source{
This object is derived from several datasets from the United States
Census Bureau, Population Division, at
\url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/counties/totals/co-est2019-alldata.pdf},
\url{https://www.census.gov/data/tables/time-series/demo/popest/2010s-total-puerto-rico-municipios.html},
and \url{https://www.census.gov/data/tables/2010/dec/2010-island-areas.html}.
It is made available through the \code{covidcast} package. This data is
public domain.
}
\usage{
state_census
}
\description{
Data set on state populations, from the 2019 US Census.
}
\section{Data dictionary}{


The data includes 57 regions (all US states, the United
States as a whole, the District of Columbia, Puerto Rico Commonwealth,
American Samoa, Guam, the U.S. Virgin Islands, and the Northern Mariana
Islands) with columns:

\describe{
\item{fips}{2-digit FIPS code}
\item{name}{Full name of the state or territory}
\item{pop}{Estimate of the location's resident population in
2019.}
\item{abbr}{Postal abbreviation for the location}
}

}

\examples{
# Since this is a re-exported dataset, it cannot be loaded using
# the `data()` function. `data()` looks for a file of the same name
# in the `data/` directory, which doesn't exist in this package.
# works
epipredict::state_census

# works
library(epipredict)
state_census

# fails
\dontrun{
data(state_census, package = "epipredict")
}
}
\keyword{datasets}
