% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_climate.R
\name{step_climate}
\alias{step_climate}
\title{Calculate a climatological variable based on the history}
\usage{
step_climate(
  recipe,
  ...,
  role = "predictor",
  time_type = c("epiweek", "week", "month", "day"),
  center_method = c("median", "mean"),
  window_size = 3L,
  epi_keys = NULL,
  prefix = "climate_",
  skip = FALSE,
  id = rand_id("climate")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? \code{lag} is default a predictor while \code{ahead} is an outcome.}

\item{time_type}{The duration over which time aggregation should be performed.}

\item{center_method}{The measure of center to be calculated over the time
window.}

\item{window_size}{Scalar integer. How many time units on each side should
be included. For example, if \code{window_size = 3} and \code{time_type = "day"},
then on each day in the data, the center will be calculated using 3 days
before and three days after. So, in this case, it operates like a weekly
rolling average, centered at each day.}

\item{epi_keys}{Character vector or \code{NULL}. Any columns mentioned will be
grouped before performing any center calculation. So for example, given
state-level data, a national climate would be calculated if \code{NULL}, but
passing \code{epi_keys = "geo_value"} would calculate the climate separately
by state.}

\item{prefix}{A character string that will be prefixed to the new column.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A unique identifier for the step}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_climate()} creates a \emph{specification} of a recipe step
that will generate one or more new columns of derived data. Using this as
a predictor is most useful when predicting strongly seasonal data. The
climate timing will automatically be aligned to the outcome at bake time.
}
\examples{
r <- epi_recipe(covid_case_death_rates) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_climate(death_rate, time_type = "day")
r

r \%>\%
  prep(covid_case_death_rates) \%>\%
  bake(new_data = NULL)
}
