% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_shift_internal.R
\name{step_epi_ahead}
\alias{step_epi_ahead}
\title{Create a shifted predictor}
\usage{
step_epi_ahead(
  recipe,
  ...,
  role = "outcome",
  trained = FALSE,
  ahead = 1,
  prefix = "ahead_",
  default = NA,
  keys = epi_keys(recipe),
  columns = NULL,
  skip = FALSE,
  id = rand_id("epi_ahead")
)
}
\arguments{
\item{shift}{A vector of integers. Each specified column will be
shifted for each value in the vector.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_epi_shift} creates a \emph{specification} of a recipe step that
will add new columns of shifted data. shifted data will
by default include NA values where the shift was induced.
These can be removed with \code{\link[=step_naomit]{step_naomit()}}, or you may
specify an alternative filler value with the \code{default}
argument.
}
\details{
The step assumes that the data are already \emph{in the proper sequential
order} for shifting.
}
\concept{row operation steps}
