% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_lag.R
\name{step_epi_lag}
\alias{step_epi_lag}
\title{Create a lagged predictor}
\usage{
step_epi_lag(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  lag = 1,
  prefix = "lag_",
  default = NA,
  keys = epi_keys(recipe),
  columns = NULL,
  skip = FALSE,
  id = rand_id("epi_lag")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from
the original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{lag}{A vector of positive integers. Each specified column will be
lagged for each value in the vector.}

\item{prefix}{A prefix for generated column names, default to "lag_".}

\item{default}{Determines what fills empty rows
left by lagging (defaults to NA).}

\item{columns}{A character string of variable names that will
be populated (eventually) by the \code{terms} argument.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_epi_lag} creates a \emph{specification} of a recipe step that
will add new columns of lagged data. Lagged data will
by default include NA values where the lag was induced.
These can be removed with \code{\link[=step_naomit]{step_naomit()}}, or you may
specify an alternative filler value with the \code{default}
argument.
}
\details{
The step assumes that the data are already \emph{in the proper sequential
order} for lagging.
}
\examples{
n <- 10
start <- as.Date("1999/01/01")
end <- as.Date("1999/01/10")

df <- data.frame(
  x = runif(n),
  index = 1:n,
  day = seq(start, end, by = "day")
)

library(recipes)
recipe(~., data = df) \%>\%
  step_lag(index, day, lag = 2:3) \%>\%
  prep(df) \%>\%
  bake(df)
}
\concept{row operation steps}
