% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_epi_naomit.R
\name{step_epi_naomit}
\alias{step_epi_naomit}
\title{Unified NA omission wrapper function for recipes}
\usage{
step_epi_naomit(recipe)
}
\arguments{
\item{recipe}{Recipe to be used for omission steps}
}
\value{
Omits NA's from both predictors and outcomes at training time;
however, only omits associated predictors at prediction time to avoid
losing data.
}
\description{
Unified NA omission wrapper function for recipes
}
\examples{
tibble(geo_value = rep("place",200),
           time_value = as.Date("2021-01-01") + 0:199,
           case_rate = 1:200,
           death_rate = 1:200) \%>\%
 as_epi_df() \%>\%
 recipe() \%>\%
 step_epi_naomit()
}
