% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_epi_naomit.R
\name{step_epi_naomit}
\alias{step_epi_naomit}
\title{Unified NA omission wrapper function for recipes}
\usage{
step_epi_naomit(recipe)
}
\arguments{
\item{recipe}{Recipe to be used for omission steps}
}
\value{
Omits NA's from both predictors and outcomes at training time
to fit the model. Also only omits associated predictors and not
outcomes at prediction time due to lack of response and avoidance
of data loss.
}
\description{
Unified NA omission wrapper function for recipes
}
\examples{
case_death_rate_subset \%>\%
 recipe() \%>\%
 step_epi_naomit()
}
