% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_shift_internal.R
\name{step_epi_lag}
\alias{step_epi_lag}
\alias{step_epi_ahead}
\title{Create a shifted predictor}
\usage{
step_epi_lag(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  lag = 1,
  default = NA,
  keys = epi_keys(recipe),
  columns = NULL,
  skip = FALSE
)

step_epi_ahead(
  recipe,
  ...,
  role = "outcome",
  trained = FALSE,
  ahead = 1,
  default = NA,
  keys = epi_keys(recipe),
  columns = NULL,
  skip = FALSE
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned?}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{lag, ahead}{A vector of nonnegative integers. Each specified column will
be the lag or lead for each value in the vector.}

\item{default}{Determines what fills empty rows
left by leading/lagging (defaults to NA).}

\item{keys}{A character vector of the keys in an epi_df}

\item{columns}{A character string of variable names that will
be populated (eventually) by the \code{terms} argument.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.

An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_epi_lag} and \code{step_epi_ahead} create a \emph{specification} of a recipe step
that will add new columns of shifted data. The former will created a lag
column, while the latter will create a lead column. Shifted data will
by default include NA values where the shift was induced.
These can be properly removed with \code{\link[=step_epi_naomit]{step_epi_naomit()}}, or you may
specify an alternative filler value with the \code{default}
argument.
}
\details{
The step assumes that the data are already \emph{in the proper sequential
order} for shifting.

The \code{prefix} and \code{id} arguments are unchangeable to ensure that the code runs
properly and to avoid inconsistency with naming. For \code{step_epi_ahead}, they
are always set to \code{"ahead_"} and \code{"epi_ahead"} respectively, while for
\code{step_epi_lag}, they are set to \code{"lag_"} and \verb{"epi_lag}, respectively.
}
\examples{
epi_recipe(case_death_rate_subset) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_lag(death_rate, lag = c(0,7,14))
}
\concept{row operation steps}
