% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_epi_shift.R
\name{step_epi_lag}
\alias{step_epi_lag}
\alias{step_epi_ahead}
\title{Create a shifted predictor}
\usage{
step_epi_lag(
  recipe,
  ...,
  lag,
  role = "predictor",
  prefix = "lag_",
  default = NA,
  skip = FALSE,
  id = rand_id("epi_lag")
)

step_epi_ahead(
  recipe,
  ...,
  ahead,
  role = "outcome",
  prefix = "ahead_",
  default = NA,
  skip = FALSE,
  id = rand_id("epi_ahead")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{lag, ahead}{A vector of integers. Each specified column will
be the lag or lead for each value in the vector. Lag integers must be
nonnegative, while ahead integers must be positive.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? \code{lag} is default a predictor while \code{ahead} is an outcome.}

\item{prefix}{A character string that will be prefixed to the new column.}

\item{default}{Determines what fills empty rows
left by leading/lagging (defaults to NA).}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A unique identifier for the step}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_epi_lag} and \code{step_epi_ahead} create a \emph{specification} of a recipe step
that will add new columns of shifted data. The \code{step_epi_lag} will created
a lagged \code{predictor} column, while \code{step_epi_ahead} will create a leading
\code{outcome} column. Shifted data will by default include NA values where the
shift was induced.  These can be properly removed with \code{\link[=step_epi_naomit]{step_epi_naomit()}},
or you may specify an alternative filler value with the \code{default} argument.
}
\details{
The step assumes that the data's \code{time_value} column is already \emph{in
the proper sequential order} for shifting.

Our \code{lag/ahead} functions respect the \code{geo_value} and \code{other_keys} of the
\code{epi_df}, and allow for discontiguous \code{time_value}s. Both of these features
are noticably lacking from \code{recipe::step_lag()}.
Our \code{lag/ahead} functions also appropriately adjust the amount of data to
avoid accidentally dropping recent predictors from the test data.

The \code{prefix} and \code{id} arguments are unchangeable to ensure that the code runs
properly and to avoid inconsistency with naming. For \code{step_epi_ahead}, they
are always set to \code{"ahead_"} and \code{"epi_ahead"} respectively, while for
\code{step_epi_lag}, they are set to \code{"lag_"} and \verb{"epi_lag}, respectively.
}
\examples{
r <- epi_recipe(covid_case_death_rates) \%>\%
  step_epi_ahead(death_rate, ahead = 7) \%>\%
  step_epi_lag(death_rate, lag = c(0, 7, 14))
r
}
\seealso{
Other row operation steps: 
\code{\link{step_adjust_latency}()},
\code{\link{step_growth_rate}()},
\code{\link{step_lag_difference}()}

Other row operation steps: 
\code{\link{step_adjust_latency}()},
\code{\link{step_growth_rate}()},
\code{\link{step_lag_difference}()}
}
\concept{row operation steps}
