% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_population_scaling.R
\name{step_population_scaling}
\alias{step_population_scaling}
\title{Create a recipe step that scales variables using population data}
\usage{
step_population_scaling(
  recipe,
  df,
  by = NULL,
  df_pop_col,
  ...,
  inputs = NULL,
  create_new = TRUE,
  suffix = "_scaled",
  role = "predictor",
  trained = FALSE,
  skip = FALSE,
  id = rand_id("population_scaling")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe. The recipe should contain information about the
\code{epi_df} such as column names.}

\item{df}{a data frame that contains the population data used for scaling.}

\item{by}{A character vector of variables to join by. Column names should be
in lower case.
To join by different variables on the \code{epi_df} and \code{df}, use a named vector.
For example, by = c("geo_value" = "states") will match \code{epi_df$geo_value}
to \code{df$states}. To join by multiple variables, use a vector with length > 1.
For example, by = c("geo_value" = "states", "county" = "county") will match
\code{epi_df$geo_value} to \code{df$states} and \code{epi_df$county} to \code{df$county}.}

\item{df_pop_col}{the name of the column in the data frame \code{df} that
contains the population data and will be used for scaling.
This should be one column, and column names should be in lower case.}

\item{...}{the corresponding column(s) in the \code{epi_df} that will be scaled.}

\item{inputs}{Quosure(s) of \code{...}.}

\item{create_new}{TRUE to create a new column and keep the original column
in the \code{epi_df}}

\item{suffix}{a character. The suffix added to the column name if
\code{crete_new = TRUE}. Default to "_scaled".}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from the
original variables will be used as predictors in a model. Other options can
be ard are not limited to "outcome".}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A unique identifier for the step}
}
\value{
Creates a population scaled column in training time to fit the model.
}
\description{
\code{step_population_scaling} creates a specification of a recipe step
that will add a population scaled column in the data. For example,
load a dataset that contains county population, and join to an \code{epi_df}
that currently only contains number of new cases by county. Once scaled,
predictions can be made on case rate.
}
