% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_window.R
\name{step_training_window}
\alias{step_training_window}
\title{Limits the size of the training window to the most recent observations}
\usage{
step_training_window(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  nrec = 50,
  skip = TRUE,
  id = rand_id("training_window")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned?}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{nrec}{An integer value that represents the number of most recent
observations that are to be kept in the training window per location.
The default value is 50.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_training_window} creates a \emph{specification} of a recipe step that
limit the size of the training window to the \code{nrec} most recent
observations in \code{time_value} per location from \code{geo_value}.
}
\examples{
tib <- tibble::tibble(
x = 1:10, y = 1:10,
time_value = rep(seq(as.Date("2020-01-01"), by = 1,
                     length.out = 5), times = 2),
geo_value = rep(c("ca", "hi"), each = 5)
) \%>\% epiprocess::as_epi_df()

library(recipes)
epi_recipe(y ~ x, data = tib) \%>\%
  step_training_window(nrec = 3) \%>\%
  prep(tib) \%>\%
  bake(new_data = NULL)
}
