% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_interval_score.R
\name{weighted_interval_score}
\alias{weighted_interval_score}
\title{Compute weighted interval score}
\usage{
weighted_interval_score(x, actual, ...)
}
\arguments{
\item{x}{dist_quantiles. A vector of class \code{\link[=dist_quantiles]{dist_quantiles()}}.}

\item{actual_value}{double. Actual value(s)}
}
\value{
a vector of nonnegative scores.
}
\description{
Weighted interval score (WIS), a well-known quantile-based
approximation of the commonly-used continuous ranked probability score
(CRPS). WIS is a proper score, and can be thought of as a distributional
generalization of absolute error. For example, see \href{https://arxiv.org/abs/2005.12881}{Bracher et al. (2020)} for discussion in the context
of COVID-19 forecasting.
}
\examples{
quantile_levels <- c(.2, .4, .6, .8)
predq_1 <- 1:4 #
predq_2 <- 8:11
dstn <- dist_quantiles(list(predq_1, predq_2), quantile_levels)
actual <- c(3.3, 7.1)
weighted_interval_score(dstn, actual)

# Using some actual forecasts --------
library(dplyr)
jhu <- case_death_rate_subset \%>\%
  filter(time_value >= "2021-10-01", time_value <= "2021-12-01")
preds <- flatline_forecaster(
    jhu, "death_rate",
    flatline_args_list(quantile_levels = c(.01, .025, 1:19 / 20, .975, .99))
)$predictions
actuals <- case_death_rate_subset \%>\%
  filter(time_value == as.Date("2021-12-01") + 7) \%>\%
  select(geo_value, time_value, actual = death_rate)
preds <- left_join(preds, actuals,
  by = c("target_date" = "time_value", "geo_value")
) \%>\%
  mutate(wis = weighted_interval_score(.pred_distn, actual))
preds
}
